% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{add_historic_data}
\alias{add_historic_data}
\title{Add historic data}
\usage{
add_historic_data(df, dates, mem.max)
}
\arguments{
\item{df}{A hop profile tibble}

\item{mem.max}{mem.max argument from \code{\link{hisafe_slice}}.}

\item{date}{date argument from \code{\link{hisafe_slice}}.}
}
\description{
Adds historic data when cells/voxels data is missing for \code{\link{hisafe_slice}}
}
