% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_phantom_trees}
\alias{add_phantom_trees}
\title{Add phantom trees}
\usage{
add_phantom_trees(tree.data)
}
\arguments{
\item{tree.data}{Tree data}
}
\description{
Determines and appends phantom trees to tree data for hisafe tile plots
}
\keyword{internal}
