% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_trees}
\alias{add_trees}
\title{Add trees & canopies to hisafe tile plots}
\usage{
add_trees(plot.obj, tree.data, trees, canopies)
}
\arguments{
\item{tree.data}{Tree data}

\item{trees}{Logical for whether or not to plot trees}

\item{canopies}{Logical for whether or not to plot canopies}

\item{hop}{An object of class hop.}
}
\description{
Adds trees & canopies to hisafe tile plots
}
\keyword{internal}
