% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roots.R
\name{addGrid}
\alias{addGrid}
\title{Add grid to hisafe_root3D() plot}
\usage{
addGrid(center, xlim, ylim, zlim, xstep, ystep, zstep)
}
\arguments{
\item{center}{An object of class hop.}

\item{xlim}{Upper & lower limits for x.}

\item{ylim}{Upper & lower limits for y.}

\item{zlim}{Upper & lower limits for z.}

\item{xstep}{Grid size in x direction.}

\item{ystep}{Grid size in y direction.}

\item{zstep}{Grid size in z direction.}
}
\description{
Adds grid to \code{\link{hisafe_root3D}} plot.
}
\author{
Gregoire Talbot \email{talbot@mailoo.org}
}
\keyword{internal}
