% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{analyze_hisafe}
\alias{analyze_hisafe}
\title{Shortcut to Hi-sAFe analysis}
\usage{
analyze_hisafe(hop, carbon = TRUE, light = TRUE, nitrogen = FALSE,
  water = FALSE, light.daily = TRUE, nitrogen.daily = FALSE,
  water.daily = FALSE, annualtree = TRUE, annualplot = TRUE,
  trees = TRUE, plot = TRUE, climate = TRUE, annualcrop = TRUE,
  monthCells = TRUE, ...)
}
\arguments{
\item{hop}{A object of class hop or face.}

\item{carbon}{Logical indicating if annual carbon plot should be made.}

\item{light}{Logical indicating if annual light plot should be made.}

\item{nitrogen}{Logical indicating if annual nitrogen plot should be made.}

\item{water}{Logical indicating if annual water plot should be made.}

\item{light.daily}{Logical indicating if daily light plots should be made.}

\item{nitrogen.daily}{Logical indicating if daily nitrogens plot should be made.}

\item{water.daily}{Logical indicating if daily water plots should be made.}

\item{annualtree}{Logical indicating if annualtree profile diagnostic plots should be made.}

\item{annualplot}{Logical indicating if annualplot profile diagnostic plots should be made.}

\item{trees}{Logical indicating if trees profile diagnostic plots should be made.}

\item{plot}{Logical indicating if plot profile diagnostic plots should be made.}

\item{climate}{Logical indicating if climate profile diagnostic plots should be made.}

\item{annualcrop}{Logical indicating if annualcrop profile diagnostic plots should be made.}

\item{monthCells}{Logical indicating if monthCells profile diagnostic plots should be made.}

\item{...}{Other arguments passed to \code{\link{plot_hisafe_ts}}.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Runs the various Hi-sAFe analysis/plot functions from a single call.
}
\examples{
\dontrun{
analyze_hisafe(myhop)
}
}
\seealso{
Other hisafe analysis functions: \code{\link{plot_hisafe_cycle}},
  \code{\link{plot_hisafe_use}}
}
