% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{annotator}
\alias{annotator}
\title{Add labels to plot panels}
\usage{
annotator(plots, labels = paste0("(", letters, ")"))
}
\arguments{
\item{plots}{A list of ggplot objects}

\item{labels}{A character vector of labels to use}
}
\description{
Adds labels to upper-right corne of plot panels
}
\keyword{internal}
