% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{arrange_hip}
\alias{arrange_hip}
\title{Arrange columns of hip object}
\usage{
arrange_hip(hip, arg.list, defaults.to.add)
}
\arguments{
\item{hip}{An object of class "hip"}

\item{arg.list}{Passed by \code{define_hisafe}.}

\item{defaults.to.add}{Passed by \code{define_hisafe}.}
}
\value{
A hip object
}
\description{
Arranges columns of hip object so that Simulationname is first, then columns that vary between eperiments,
then columns with non-default values, and finally columns with default values.
Used within \code{define_hisafe} and \code{define_hisafe_file}.
}
