% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face.R
\name{augment_with_system}
\alias{augment_with_system}
\title{Augment hop elements with system name}
\usage{
augment_with_system(hop, system)
}
\arguments{
\item{hop}{An object of class "hop" to augment}

\item{system}{A character string of the system name to use}
}
\value{
Returns the modified hop object
}
\description{
Adds a "System" column to all hop elements
}
