% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{build_cluster_script}
\alias{build_cluster_script}
\title{Builds a cluster shell script}
\usage{
build_cluster_script(simu.name, script.path, cluster.path, email,
  num.cores = length(simu.name))
}
\arguments{
\item{simu.name}{A character string of the name of the simulation to create the script for.
This must match the name of .SIM file as well as the name of the folder that file is in.}

\item{script.path}{A character string of the local path for where to create the cluster script file.}

\item{cluster.path}{A character string of the path on the cluster where the simulation folder is located.}

\item{email}{A character string of the email address to notify with cluster job updates.}

\item{num.cores}{The number of cores to request from the cluster.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Builds a shell script to initialize Hi-sAFe simulations on the MUSE cluster.
}
\examples{
\dontrun{
mysim <- build_cluster_script("mysim",
                              script.path  = "/simulations/",
                              cluster.path = "/nfs/work/hisafe/simulations/user/",
                              email        = "me@work.com")
}
}
