% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_hisafe_exp}
\alias{build_hisafe_exp}
\title{Builds a group Hi-sAFe simulations}
\usage{
build_hisafe_exp(hip, path, exp.name = "experiment",
  export.profiles = "all", saveProjectOption = FALSE, controls = FALSE)
}
\arguments{
\item{hip}{An object of class hip - typically with multiple experiments (rows).}

\item{path}{A character string of the path to the folder in which the experiment folder should be created.}

\item{export.profiles}{A character vector of output profiles the simulation to export.}

\item{saveProjectOption}{Logical, sets the saveProjectOption parameter in the .sim file.}
}
\value{
Invisibly returns a list containing the original hip object and path to the experiment folder.
}
\description{
Builds a Hi-sAFe experiment (a group of simulation) - creates the folder structure and input files.
}
\examples{
\dontrun{
# Once a group Hi-sAFe simulations (experiment) is defined:
myexp <- define_hisafe(latitude = c(30,60))

# Building the experiment folder structure & files:
build_hisafe_exp(myexp, "./simulations", exp.name = "lat_exp")
}
}
\seealso{
Other hisafe build functions: \code{\link{build_hisafe}}
}
