% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_pld}
\alias{build_pld}
\title{Build .pld file for Hi-sAFe simulation}
\usage{
build_pld(plan, path)
}
\arguments{
\item{plan}{A single simulation (row) from an object of class "hip"}

\item{path}{A character string of the path to the simulation folder.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Builds .pld file for a Hi-sAFe simulation. Used within \code{build_hisafe}.
}
