% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_sim}
\alias{build_sim}
\title{Build .sim file for Hi-sAFe simulation}
\usage{
build_sim(plan, path, profiles, saveProjectOption)
}
\arguments{
\item{plan}{A single simulation (row) from an object of class hip.}

\item{path}{A character string of the path to the simulation folder.}

\item{profiles}{A character vector of output profiles the simulation to export.}

\item{saveProjectOption}{Logical, sets the saveProjectOption parameter in the .sim file.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Builds .sim file for a Hi-sAFe simulation. Used within \code{build_hisafe}.
}
