% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_structure}
\alias{build_structure}
\title{Builds a Hi-sAFe simulation}
\usage{
build_structure(hip, path, profiles, saveProjectOption)
}
\arguments{
\item{hip}{An object of class "hip" containing a single simulation (row).}

\item{path}{A character string of the path to the simulation folder.}

\item{profiles}{A character vector of output profiles the simulation to export.}

\item{saveProjectOption}{Logical, sets the saveProjectOption parameter in the .sim file.}
}
\value{
Invisibly returns a list containing the original hip object and supplied path.
}
\description{
Builds a Hi-sAFe simulation - creates the folder structure and input files.
}
