% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_tree}
\alias{build_tree}
\title{Edit .tree file for Hi-sAFe simulation}
\usage{
build_tree(plan, path, species)
}
\arguments{
\item{plan}{A single simulation (row) from an object of class "hip"}

\item{path}{A character string of the path to the simulation folder.}

\item{species}{A character string of the tree species to edit.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Edits .tree file for a Hi-sAFe simulation. Used within \code{build_hisafe}.
}
