% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{build_white_boxes_tile}
\alias{build_white_boxes_tile}
\title{Build white boxes to cover phantom trees}
\usage{
build_white_boxes_tile(hop, X.MIN, X.MAX, Y.MIN, Y.MAX)
}
\arguments{
\item{hop}{An object of class hop.}

\item{X.MIN}{Lower x limit for plot.}

\item{X.MAX}{Upper x limit for plot.}

\item{Y.MIN}{Lower y limit for plot.}

\item{Y.MAX}{Upper y limit for plot.}
}
\description{
Builds white boxes to cover phantom trees for hisafe tile plot functions
}
