% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{call_hisafe}
\alias{call_hisafe}
\title{Calls Hi-sAFe from command line}
\usage{
call_hisafe(path, simu.name, capsis.path, launch.call, default.folder,
  quietly)
}
\arguments{
\item{path}{A character string of the path to the folder containing the simulation folder. Required if \code{hip} is not provided.}

\item{simu.name}{Name of the simulation to run. Required if \code{hip} is not provided.}

\item{capsis.path}{A character string of the path to the Capsis folder}

\item{quietly}{Logical indicating whether status messages should printed to the console.}
}
\value{
Invisibly returns a console log from simulation run.
}
\description{
Calls Hi-sAFe from command line to run a simulation.
}
\keyword{internal}
