% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{check_allowed}
\alias{check_allowed}
\title{Check validity of Hi-sAFe input ranges}
\usage{
check_allowed(variable, hip)
}
\arguments{
\item{variable}{A character string of the name of the variable to check.}

\item{hip}{An object of class "hip"}
}
\value{
An error message or empty character stirng.
}
\description{
Checks validity of Hi-sAFe inputs ranges against ranges in \code{HISAFE.PARAMS}.
Used within \code{check_input_values}.
}
