% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{check_input_values}
\alias{check_input_values}
\title{Check validity of all Hi-sAFe inputs}
\usage{
check_input_values(hip)
}
\arguments{
\item{hip}{An object of class \code{hip}.}
}
\value{
Produces errors and/or warnings if issues are found. Otherwise, invisibly returns \code{TRUE}.
}
\description{
Checks the validity of all Hi-sAFe inputs against \code{HISAFE.PARAMS} and provides errors and/or warnings if issues are found.
Used within \code{define_hisafe} and \code{define_hisafe_file}.
}
