% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{check_minmax}
\alias{check_minmax}
\title{Check validity of Hi-sAFe input ranges}
\usage{
check_minmax(variable, hip)
}
\arguments{
\item{variable}{A character string of the name of the variable to check.}

\item{hip}{An object of class \code{hip}.}
}
\value{
An error message or empty character stirng.
}
\description{
Checks validity of Hi-sAFe input min and max values against values in \code{HISAFE.PARAMS}.
Used within \code{check_input_values}.
}
