% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{check_minmax_sug}
\alias{check_minmax_sug}
\title{Check plausbility of Hi-sAFe input ranges}
\usage{
check_minmax_sug(variable, hip)
}
\arguments{
\item{variable}{A character string of the name of the variable to check.}

\item{hip}{An object of class "hip"}
}
\value{
A warning message or empty character stirng.
}
\description{
Checks plausibility of Hi-sAFe input min and max values against suggested values in \code{HISAFE.PARAMS}.
Used within \code{check_input_values}.
}
