% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{check_type}
\alias{check_type}
\title{Check validity of Hi-sAFe input numeric types}
\usage{
check_type(variable, exp.plan, template)
}
\arguments{
\item{variable}{A character string of the name of the variable to check.}

\item{exp.plan}{The exp.plan of a "hip" object.}

\item{template}{A list of template parameters and constraints}
}
\value{
An error message or empty character stirng.
}
\description{
Checks validity of Hi-sAFe inputs against accepted numeric types (continuous, integer) found in the comments of template files.
Used within \code{\link{check_input_values}}.
}
