% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{compress_batch_array}
\alias{compress_batch_array}
\title{Compress a batch array into a character string}
\usage{
compress_batch_array(x)
}
\arguments{
\item{x}{A numeric vector containing the batch array values.}
}
\value{
A character string with continuous sequences seaprated by "-" and non-continuous sequences separated by ",".
}
\description{
Compresses a batch array into a character string
}
\keyword{internal}
