% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face.R
\name{create_face}
\alias{create_face}
\title{Combine hop objects into a face object}
\usage{
create_face(agroforestry, forestry, monocrop, face.path)
}
\arguments{
\item{agroforestry}{An object of class "hop" containing one or more agroforestry simulations.}

\item{forestry}{An object of class "hop" containing a single simulation to serve as a forestry control.}

\item{monocrop}{An object of class "hop" containing a single simulation to serve as a monocrop control. This simulation must not contain any trees.}

\item{face.path}{A character string of the path to a directory where any face analyses should be written to.}
}
\value{
An object of class "face", which is a list of three objects of class "hop":
\itemize{
 \item{agroforestry}
 \item{forestry}
 \item{monocrop}
}
}
\description{
Combines three hop objects for agroforestry, forestry, and monocrop simulations into a face object.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulations for agroforestry, forestry, and monocrop simulations via:
AF.hop <- read_hisafe(myAFhip)
FC.hop <- read_hisafe(myFChip)
CC.hop <- read_hisafe(myCChip)

# Combine these into a "face" object via:
my.face <- create_face(AF.hop, FC.hop, CC.hop)
}
}
