% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_tree_data}
\alias{create_tree_data}
\title{Generate tree data for hisafe tile plots}
\usage{
create_tree_data(hop, trees, canopies, plot.x)
}
\arguments{
\item{hop}{An object of class hop.}

\item{trees}{Logical for whether or not to plot trees}

\item{canopies}{Logical for whether or not to plot canopies}

\item{plot.x}{The plot.x argument from the tile plot function}
}
\description{
Generates tree data for hisafe tile plots
}
\keyword{internal}
