% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{cycle_budget}
\alias{cycle_budget}
\title{Calculate water and nitrogen budgets}
\usage{
cycle_budget(hop, cycle, simu.names = "all", years = "all",
  save.table = TRUE, save.plot = TRUE, output.path = NULL)
}
\arguments{
\item{hop}{An object of class hop or face.}

\item{cycle}{A character vector of the cycle to analyze Supported cycles include: 'water', and 'nitrogen.}

\item{simu.names}{A character vector of the SimulationNames within \code{hop} to include. Use "all" to include all available values.}

\item{years}{A numeric vector of the years within \code{hop} to include. Use "all" to include all available values.}

\item{save.table}{Logical indicating whether a table of budget data and calculations should be saved to \code{output.path}.}

\item{save.plot}{Logical indicating whether a plot of the budget excess should be saved to \code{output.path}.}

\item{output.path}{A character string indicating the path to the directory where plots should be saved.
If \code{NULL}, the experiment/simulation path is read from the hop object, and a directory is created there called "analysis/budgets".
The tables and plots will be saved in this directory.}
}
\value{
A data.frame (tibble) containing the budget data.
}
\description{
Calculates water and nitrogen budgets
See \code{\link{plot_hisafe_cycle_annual}} for details of each cycle.
}
\examples{
\dontrun{
water.budget <- cycle_budget(myhop, "water")
}
}
\seealso{
Other hisafe analysis functions: \code{\link{LER_summary}},
  \code{\link{LER}}, \code{\link{analyze_hisafe}},
  \code{\link{cycle_summary}},
  \code{\link{plot_hisafe_cycle_annual}},
  \code{\link{plot_hisafe_cycle_daily}},
  \code{\link{write_hop}}
}
\concept{hisafe analysis functions}
