% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{cycle_summary}
\alias{cycle_summary}
\title{Plot summary dashboard of daily and annual cycles}
\usage{
cycle_summary(hop, daily.year, simu.name = NULL, cycles = c("carbon",
  "light", "water", "nitrogen"), crop.names = c("Main crop",
  "Inter crop"), output.path = NULL, plot.labels = NULL)
}
\arguments{
\item{hop}{An object of class hop or face.}

\item{simu.name}{A character vector of legnth 1 indicating the SimulationName within \code{hop} to plot.}

\item{cycles}{A character vector of the cycles to include. Supported cycles include: 'carbon', 'light', 'water', and 'nitrogen.
See \code{\link{plot_hisafe_cycle_annual}} for details of each cycle.}

\item{crop.names}{A character vector of length 2 containing the names to use in the legend for the mainCrop and interCrop of Hi-sAFe, in that order.}

\item{output.path}{A character string indicating the path to the directory where plots should be saved.
If \code{NULL}, the experiment/simulation path is read from the hop object, and a directory is created there called "analysis".
The plot wil be saved in this directory as "cycles_summary_SimulationName.jpg".}

\item{plot.labels}{A character vector of labels to label upper-right corner of plot panels. If \code{NULL}, no labels will be added.}

\item{daily.years}{A numeric vector of legnth 1 indicating the year to use for the daily plots.}
}
\value{
Invisibly retruns an egg object.
}
\description{
Plots summary dashboard of daily and annual cycles.
}
\examples{
\dontrun{
cycle.plot <- cycle_summary(myhop, 2000)
}
}
\seealso{
Other hisafe analysis functions: \code{\link{LER_summary}},
  \code{\link{LER}}, \code{\link{analyze_hisafe}},
  \code{\link{cycle_budget}},
  \code{\link{plot_hisafe_cycle_annual}},
  \code{\link{plot_hisafe_cycle_daily}},
  \code{\link{write_hop}}
}
\concept{hisafe analysis functions}
