% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define_hisafe_file}
\alias{define_hisafe_file}
\title{Define a Hi-sAFe experiment from a file}
\usage{
define_hisafe_file(file)
}
\arguments{
\item{file}{A character string of the path to a csv file.
Each row in the file should represent a Hi-sAFe simulation and each column a Hi-sAFe input parameter.
For more information on supported parameters, use \code{hisafe_params()}.}
}
\value{
An object of class "hip". This is a data frame (tibble) with each row a Hi-sAFe simulation and each column a Hi-sAFe input parameter.
}
\description{
Defines a Hi-sAFe experiment - the input parameters to one or more Hi-sAFe simulations.
}
\details{
It is strongly recommended to name each simulation in your experiment. This can be done via the \code{SimulationName} parameter.
If no names are provided, then \code{define_hisafe_file} will provide generic names of "Sim_1", "Sim_2", etc.
}
\examples{
\dontrun{
# To define a Hi-sAFe experiment from a file:
myexp <- define_hisafe_file("./data/example_exp.csv")
}
}
\seealso{
Other hisafe definition functions: \code{\link{define_hisafe}}
}
