% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diag_hisafe}
\alias{diag_hisafe}
\title{Shortcut to Hi-sAFe diagnostics}
\usage{
diag_hisafe(hop, annualTrees = FALSE, annualPlot = FALSE, trees = TRUE,
  plot = TRUE, climate = TRUE, annualCells = FALSE, monthCells = TRUE,
  cells = TRUE, voxels = TRUE, ...)
}
\arguments{
\item{hop}{An object of class hop or face.}

\item{annualTrees}{Logical indicating if annualTrees profile diagnostic plots should be made.}

\item{annualPlot}{Logical indicating if annualPlot profile diagnostic plots should be made.}

\item{trees}{Logical indicating if trees profile diagnostic plots should be made.}

\item{plot}{Logical indicating if plot profile diagnostic plots should be made.}

\item{climate}{Logical indicating if climate profile diagnostic plots should be made.}

\item{annualCells}{Logical indicating if annualCells profile diagnostic plots should be made.}

\item{monthCells}{Logical indicating if monthCells profile diagnostic plots should be made.}

\item{voxels}{Logical indicating if voxels profile diagnostic plots should be made.}

\item{...}{Other arguments passed to \code{\link{plot_hisafe_ts}}.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Runs the various Hi-sAFe diagnostic functions from a single call.
}
\examples{
\dontrun{
diag_hisafe(myhop)
}
}
