% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diag_hisafe_annualcells}
\alias{diag_hisafe_annualcells}
\title{Plot annualCells diagnostics of Hi-sAFe output}
\usage{
diag_hisafe_annualcells(hop, years = seq(min(hop$annualCells$Year),
  max(hop$annualCells$Year), 5), output.path = NULL, ...)
}
\arguments{
\item{hop}{An object of class "hop" or "face" containing output data from one or more Hi-sAFe simulations.}

\item{years}{A numeric vector containing the years (after planting) to include. Use "all" to include all available values.}

\item{output.path}{A character stting indicating the path to the directory where plots should be saved. Plots are
saved in a subdirectory within this directory named /annualCells
If no value is provided, the experiment/simulation path is read from the hop object, and a directory is created there called "analysis/diagnostics".}

\item{...}{Other arguments passed to \code{\link{plot_hisafe_annualcells}}.}
}
\value{
Invisibly returns a list of \code{ggplot} objects.
}
\description{
Creates a tile plot of every Hi-sAFe annualCells output variable, with one plot for every 5 simulation years.
All plots are saved as png files to a specifified output path.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe(path = "mydir", simu.names = "MySimulation")

# You can create tile plots of every annualCells variable:
diag_hisafe_annualcells(mydata)
}
}
\seealso{
Other hisafe diagnostic fucntions: \code{\link{diag_hisafe_cells}},
  \code{\link{diag_hisafe_monthcells}},
  \code{\link{diag_hisafe_ts}},
  \code{\link{diag_hisafe_voxels}}
}
\concept{hisafe diagnostic fucntions}
