% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diag_hisafe_monthcells}
\alias{diag_hisafe_monthcells}
\title{Plot monthCells diagnostics of Hi-sAFe output}
\usage{
diag_hisafe_monthcells(hop, output.path = NULL, schemes = c("year.simname",
  "month.simname", "month.year"), trees = TRUE, canopies = TRUE)
}
\arguments{
\item{hop}{An object of class "hop" or "face" containing output data from one or more Hi-sAFe simulations.}

\item{output.path}{A character stting indicating the path to the directory where plots should be saved. Plots are
saved in a subdirectory within this directory named /monthCells/facetScheme.
If no value is provided, the experiment/simulation path is read from the hop object, and a directory is created there called "analysis/diagnostics".}

\item{schemes}{A character vector of the facet schemes to run. Possible schemes are described in Details and have the following names: \code{year.simname}, \code{month.simname}, \code{month.year}.}

\item{trees}{Logical indicating if a point should be plotted at the location of each tree.}

\item{canopies}{Logical indicating if an elipsoid should be plotted representing the size of each tree canopy.}
}
\value{
Invisibly returns a list of \code{ggplot} objects, grouped into two sublists based on the faceting scheme.
The first sublist contains  Year~SimulationName, and the second contains Month~SimulationName.
}
\description{
Creates three tile plots of every Hi-sAFe monthCells output variable, one plot for each possible
\code{plot_hisafe_monthcells} facet scheme. All plots are saved as png files to a specifified output path.
}
\details{
The default data presentation for each of the facet schemes is:
Year~SimulationName plots are for Month==6, every 5 years from 0:max, and all simulations.
Month~SimulationName plots are for Year==median, all 12 months, and all simulations.
Month~Year plots are for all 12 months and every 5 years from 0:max, with each simulation in its own plot.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe(path = "mydir", simu.names = "MySimulation")

# You can create tile plots of every monthCells variable:
diag_hisafe_monthcells(mydata)
}
}
\seealso{
Other hisafe diagnostic fucntions: \code{\link{diag_hisafe_annualcrop}},
  \code{\link{diag_hisafe_ts}},
  \code{\link{diag_hisafe_voxels}}
}
