% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{diag_hisafe_ts}
\alias{diag_hisafe_ts}
\title{Plot timeseries diagnostics of Hi-sAFe output}
\usage{
diag_hisafe_ts(data, time.class = "Annual", output.path = "./diagnostics")
}
\arguments{
\item{data}{An object of class \code{hisafe} or \code{hisafe-group} containing output data from one or more Hi-sAFe simulations.}

\item{time.class}{If 'Annual', annual timeseries are created. If 'Daily', daily timeseries are created.}

\item{output.path}{A character stting indicating the path to the directory where plots should be saved. Plots are
saved in a subdirectory within this directory named by \code{time.class}.}
}
\value{
A list of \code{ggplot} objects is invisibly returned. If the data is of class \code{hisafe-group} and contains
data from more than one Hi-sAFe simulation, the plots will contain multiple lines, colored and labeled by SimName.
If the data contains two more tree ids, the plots will be faceted by tree id.
}
\description{
Plots a daily or annual timeseries of every Hi-sAFe output variable. All plots are saved as
png files to a specifified output path.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe_output("MySimulation", "./")

# You can create an annual timeseries of every variable:
diag_hisafe_ts(mydata)

# For aily timeseries instead:
diag_hisafe_ts(mydata, "Daily")
}
}
