% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diag_hisafe_voxels}
\alias{diag_hisafe_voxels}
\title{Plot timeseries diagnostics of Hi-sAFe voxels output}
\usage{
diag_hisafe_voxels(hop, output.path = NULL, ...)
}
\arguments{
\item{hop}{An object of class "hop" or "face" containing output data from one or more Hi-sAFe simulations.}

\item{output.path}{A character stting indicating the path to the directory where plots should be saved.
Plots aresaved in a subdirectory within this directory named by \code{profile}.
If no value is provided, the experiment/simulation path is read from the hop object, and a directory is created there called "analysis/diagnostics".}

\item{...}{Other arguments passed to \code{\link{plot_hisafe_voxels}}.}
}
\value{
Invisibly returns a list of \code{ggplot} objects. If \code{hop} contains
data from more than one Hi-sAFe simulation, the plots will by faceted by SimulationName.
}
\description{
Plots a dail timeseries of every Hi-sAFe output variable in the voxels output profile.
All plots are saved as png files to a specifified output path.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe(path = "mydir", simu.names = "MySimulation")

# You can create an annual timeseries of every variable:
diag_hisafe_ts(mydata)

# For daily timeseries instead:
diag_hisafe_ts(mydata, "daily")
}
}
\seealso{
Other hisafe diagnostic fucntions: \code{\link{diag_hisafe_annualcrop}},
  \code{\link{diag_hisafe_monthcells}},
  \code{\link{diag_hisafe_ts}}
}
