% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diag_output_path}
\alias{diag_output_path}
\title{Determine diagnostic output path}
\usage{
diag_output_path(hop, output.path)
}
\arguments{
\item{hop}{An object of class hop or face}

\item{output.path}{The output.path argument from a diagnostic function}
}
\value{
The output path
}
\description{
Determines diagnostic output path
}
\keyword{internal}
