% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{edit_param_element}
\alias{edit_param_element}
\title{Edit an individual parameter element}
\usage{
edit_param_element(param.list, variable, value)
}
\arguments{
\item{param.list}{A list containing all parameter values and constraints.}

\item{variable}{A character string of the name of the variable to edit.}

\item{value}{A numeric vector of the value to apply to the variable.}
}
\value{
A edited list containing all parameter values and constraints.
}
\description{
Edits an individual element of a parameter list
Used within \code{\link{build_structure}}.
}
\seealso{
Other hisafe param functions: 
\code{\link{edit_param_file}()},
\code{\link{read_param_file}()},
\code{\link{write_param_file}()}
}
\concept{hisafe param functions}
\keyword{internal}
