% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{edit_param_file}
\alias{edit_param_file}
\title{Edits parameter files}
\usage{
edit_param_file(param.list, exp.plan)
}
\arguments{
\item{param.list}{A list containing all parameter values and constraints.}

\item{exp.plan}{A exp.plan element of a hip object.}
}
\value{
A edited list containing all parameter values and constraints.
}
\description{
Edits parameter files using the exp.plan of a hip object
Used within \code{\link{build_structure}}.
}
\seealso{
Other hisafe param functions: \code{\link{edit_param_element}},
  \code{\link{read_param_file}},
  \code{\link{write_param_file}}
}
\concept{hisafe param functions}
\keyword{internal}
