% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{get_carbon_allocation}
\alias{get_carbon_allocation}
\title{Get tree carbon above vs. belowground allocation from a hop object}
\usage{
get_carbon_allocation(hop)
}
\arguments{
\item{hop}{An object of class hop or face.}
}
\value{
A tibble with extracted and calculated carbon above vs. belowground allocation.
}
\description{
Gets tree carbon above vs. belowground allocation from a hop object.
Used within hisafe cycle functions.
}
\keyword{internal}
