% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{get_carbon_increment}
\alias{get_carbon_increment}
\title{Get tree carbon increment from a hop object}
\usage{
get_carbon_increment(hop)
}
\arguments{
\item{hop}{An object of class hop or face.}
}
\value{
A tibble with extracted and calculated carbon increments.
}
\description{
Get tree carbon increment from a hop object.
Used within hisafe cycle functions.
}
\keyword{internal}
