% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{get_carbon_pools}
\alias{get_carbon_pools}
\title{Get carbon pools from a hop object}
\usage{
get_carbon_pools(hop)
}
\arguments{
\item{hop}{An object of class hop or face.}
}
\value{
A tibble with extracted and calculated carbon pools.
}
\description{
Get carbon pools from a hop object.
Used within \code{\link{plot_hisafe_cycle}}, \code{\link{plot_hisafe_cycle_annual}}, \code{\link{plot_hisafe_cycle_daily}}
}
