% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{get_nitrogen_fluxes}
\alias{get_nitrogen_fluxes}
\title{Get nitrogen fluxes from a hop object}
\usage{
get_nitrogen_fluxes(hop)
}
\arguments{
\item{hop}{An object of class hop or face.}
}
\value{
A tibble with extracted and calculated nitrogen fluxes.
}
\description{
Get nitrogen fluxes from a hop object.
Used within \code{\link{plot_hisafe_cycle}}, \code{\link{plot_hisafe_cycle_annual}}, \code{\link{plot_hisafe_cycle_daily}}
}
