% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{get_param_names}
\alias{get_param_names}
\title{Get names of template parameters from sim, pld, and tree files}
\usage{
get_param_names(x)
}
\arguments{
\item{x}{A list containing all parameter values and constraints.}
}
\value{
A list containing parameter names by file type.
}
\description{
Get names of all template parameters from sim, pld, and tree files
}
