% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{get_param_vals}
\alias{get_param_vals}
\title{Get values/constraints of template parameters}
\usage{
get_param_vals(x, type)
}
\arguments{
\item{x}{A list containing all parameter values and constraints.}
}
\value{
A list containing parameter values/constraints.
}
\description{
Gets values/constraints of all template parameters
}
\keyword{internal}
