% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_pruning_dates}
\alias{get_pruning_dates}
\title{Get branch pruning dates from a hop}
\usage{
get_pruning_dates(hop, type = "branch")
}
\arguments{
\item{hop}{An object of class hop or face.}

\item{type}{One of "branch" or "root".}
}
\value{
A tibble (data.frame)
}
\description{
Gets branch pruning dates from a hop in a format suitable for supplementing plots.
}
\examples{
\dontrun{
pruning.dates <- get_pruning_dates(hop)
}
}
\seealso{
Other hisafe helper functions: \code{\link{copy_hisafe_template}},
  \code{\link{get_pheno_dates}}, \code{\link{hip_params}},
  \code{\link{hisafe_info}}, \code{\link{hisafe_profiles}},
  \code{\link{hop_filter}}, \code{\link{hop_merge}},
  \code{\link{hop_params}}, \code{\link{hop_rename}},
  \code{\link{is_face}}, \code{\link{is_hip}},
  \code{\link{is_hop}}, \code{\link{join_profiles}},
  \code{\link{profile_check}},
  \code{\link{variable_check}},
  \code{\link{which_profiles}}
}
\concept{hisafe helper functions}
