% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\name{get_stics_budget}
\alias{get_stics_budget}
\title{Get stics budget values from .bil export files}
\usage{
get_stics_budget(year, simu.name, path, stics.names, start.header, end.header)
}
\arguments{
\item{year}{Year of the file to extract.}

\item{simu.name}{Year of the file to extract.}

\item{path}{Path to simulation folder.}

\item{stics.names}{Names of the stics budget variables.}

\item{start.header}{Character string of the starting line to search.}

\item{end.header}{Character string of the final line to search.}
}
\value{
A tibble with extracted values.
}
\description{
Gets stics budget values from .bil export files.
Used within \code{\link{stics_budget_comp}} functions.
}
\keyword{internal}
