% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_units}
\alias{get_units}
\title{Get variable units}
\usage{
get_units(variable, prof)
}
\arguments{
\item{variable}{A character string of the name of the variable to get units for.}

\item{prof}{A characgter string of the profile that \code{variable} is from.}
}
\description{
Gets variable units. Used within all hisafe plot functions.
}
\keyword{internal}
