% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{get_used_param}
\alias{get_used_param}
\title{Gets the parameter actually used}
\usage{
get_used_param(variable, exp.plan, template.defaults, template.commented)
}
\arguments{
\item{variable}{A character string of the name of the variable to check.}

\item{template.defaults}{A list containing the default values for all parameters.}

\item{hip}{A "hip" object}
}
\value{
The value of the parameter
}
\description{
Gets the parameter actually used (i.e. the default or the defined)
}
