% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{get_used_params}
\alias{get_used_params}
\title{Complies list of parameters actually used}
\usage{
get_used_params(hip)
}
\arguments{
\item{hip}{A "hip" object containing only a single simulation}
}
\value{
List of used parameter values
}
\description{
Compiles list of the parameters actually used (i.e. the default or the defined)
}
