% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{get_yields}
\alias{get_yields}
\title{Get yields from a hop object}
\usage{
get_yields(hop, profile, crop.names)
}
\arguments{
\item{hop}{An object of class hop or face.}

\item{profile}{An character string indicating from which profile to pull crop yield data. Either "cells" or "plot".}

\item{crop.names}{A character vector of length 2 containing the names to use in the legend for the mainCrop and interCrop of Hi-sAFe, in that order.}
}
\value{
A tibble with extracted yields.
}
\description{
Gets yields from a hop object.
Used within hisafe cycle functions.
}
\keyword{internal}
