% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hip_params}
\alias{hip_params}
\title{Display supported Hi-sAFe input parameters}
\usage{
hip_params(params = "all", search = FALSE, template = "agroforestry")
}
\arguments{
\item{params}{A character vector of specific Hi-sAFe input parameters of which to display details.
Can also be a regular expression for which to search in Hi-sAFe parameter names.
If "all", the default, then a data.frame (tibble) containing all Hi-sAFe input parameters is returned.}

\item{search}{Logical indicating whether \code{params} should be treated as a regular expression and
searched for in the parameter names rather than matched literally.}

\item{template}{A character string of the path to the directory containing the template set of Hi-sAFe simulation
folders/files to use.
hisafer comes with a variety of "default" templates than can be used by specificying specific character strings.
See \code{\link{define_hisafe}} for details.}
}
\value{
If \code{params} is "all", the default, then a data.frame (tibble) containing all Hi-sAFe input parameters
is returned. Otherwise, this data.frame is invisibly returned.
}
\description{
Displays supported Hi-sAFe input parameters, their default values, and their accepted/suggested ranges.
}
\examples{
\dontrun{
hip_params()              # details of all parameters
hip_params("cellWidth")   # details of cellWidth parameter
hip_params("paramShape1") # input parameters within tables are also accepted
}
}
\seealso{
Other hisafe helper functions: 
\code{\link{copy_hisafe_template}()},
\code{\link{get_pheno_dates}()},
\code{\link{get_pruning_dates}()},
\code{\link{hisafe_info}()},
\code{\link{hisafe_profiles}()},
\code{\link{hop_filter}()},
\code{\link{hop_merge}()},
\code{\link{hop_params}()},
\code{\link{hop_rename}()},
\code{\link{is_face}()},
\code{\link{is_hip}()},
\code{\link{is_hop}()},
\code{\link{join_profiles}()},
\code{\link{make_rel_years}()},
\code{\link{profile_check}()},
\code{\link{variable_check}()},
\code{\link{which_profiles}()}
}
\concept{hisafe helper functions}
