% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hisafe_info}
\alias{hisafe_info}
\title{Display version numbers of Hi-sAFe and Java}
\usage{
hisafe_info(capsis.path)
}
\arguments{
\item{capsis.path}{A character string of the path to the Capsis folder}
}
\value{
Invisibly returns the Hi-sAFe version number
}
\description{
Displays the version numbers of Hi-sAFe and Java.
}
\examples{
\dontrun{
hisafe_info()
}
}
