% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hisafe_params}
\alias{hisafe_params}
\title{Display supported Hi-sAFe input parameters}
\usage{
hisafe_params(variable = "names")
}
\arguments{
\item{variable}{If "names", the default, then just the names of supported Hi-sAFe parameters is printed to the console.
If "all", then the names, default values, and accepted/suggested ranges of supported Hi-sAFe parameters is printed.
Can also be a character vector of specific Hi-sAFe parameters of which to display details.}
}
\value{
Invisibly returns an alphebetized character vector of the names of supported Hi-sAFe prameters.
}
\description{
Displays supported Hi-sAFe input parameters, their default values, and their accepted/suggested ranges.
}
\examples{
\dontrun{
hisafe_params()            # just for parameter names
hisafe_params("cellWidth") # details of cellWidth parameter
hisafe_params("all")       # details of all parameters
}
}
\seealso{
Other hisafe definition functions: \code{\link{define_hisafe_file}},
  \code{\link{define_hisafe}},
  \code{\link{hisafe_profiles}}
}
