% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roots.R
\name{hisafe_root3D}
\alias{hisafe_root3D}
\title{Create a 3D plot of tree coarse root topology}
\usage{
hisafe_root3D(hop, date, simu.name = NULL, tree.id = 1,
  color.var = paste0("treeRootDensity_", tree.id),
  color.palette = c("#FFFFD4", "#FED98E", "#FE9929", "#D95F0E",
  "#993404"), bg = "white", grid = FALSE, box = FALSE,
  output.path = NULL, ...)
}
\arguments{
\item{hop}{An object of class hop.}

\item{date}{A character string of the date to plot, in the format "YYYY-MM-DD" or of class Date.}

\item{simu.name}{A character vector of legnth 1 indicating the SimulationName within \code{hop} to plot.}

\item{tree.id}{A numeric vector of lenght 1 indicating the tree id to plot.}

\item{color.var}{A character string indicating the name of the voxels variable to use for coloring the roots.}

\item{color.palette}{A character vector of colors for coloring the roots to pass to \code{grDevices::colorRampPalette}.}

\item{bg}{A character string of hex value or R standard color name defining the plot background color.}

\item{grid}{A logical indicating whether the plot should include an x,y,z grid.}

\item{box}{A logical indicating whether the plot should include a boundign box.}

\item{output.path}{A character string indicating the path to the directory where plots should be saved.
If \code{NULL}, the experiment/simulation path is read from the hop object, and a directory is created there called "analysis/roots3D".
The plot will be saved in this directory as "simu.name_date_root3D.png".}

\item{...}{Other arguments passed to \code{rgl::par3d}.}
}
\description{
Creates a 3D plot of tree coarse root topology. 
The plot is saved as soon as rgl has finished drawing it on the screen. 
The default view is typically not very good, so zooming/rotating the plot while it's drawing can allow you to save a better plot.
\code{rgl::rgl.snapshot} can also be used directly to save the plot once the desired view is found.
}
\details{
Key for voxel output "colonisationDirection":
\itemize{
 \item{0}{: x-}
 \item{1}{: x-}
 \item{2}{: y+}
 \item{3}{: y-}
 \item{4}{: z+}
 \item{5}{: z-}
}
Positive z (+z) is in the downward direction. 
Everything above specifies the direction from the parent voxel to the child voxel and is stored in the child.
}
\examples{
\dontrun{
root3D(hop = myhop, date = "2000-07-01")
}
}
\seealso{
Other hisafe visualization functions: \code{\link{hisafe_slice}},
  \code{\link{hisafe_snapshot}}
}
\author{
Gregoire Talbot \email{talbot@mailoo.org}

Kevin Wolz
}
\concept{hisafe visualization functions}
