% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisafer.R
\docType{package}
\name{hisafer}
\alias{hisafer}
\alias{hisafer-package}
\title{hisafe: A package of tools for interacting with the Hi-sAFe biophysical agroforestry model.}
\description{
The hisafer package provides functions for defining, building, running, reading,
analyzing, and plotting Hi-sAFe simulaitons.
}
