% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hop_date_filter}
\alias{hop_date_filter}
\title{Filter a hop object by date}
\usage{
hop_date_filter(hop, date.min = NA, date.max = NA)
}
\arguments{
\item{hop}{A object of class hop or face.}

\item{date.min}{A character string of the minimum date to keep, in the format "YYYY-MM-DD".
If NA, the minimum date in \code{hop} is used.}

\item{date.max}{A character string of the maximum date to keep, in the format "YYYY-MM-DD".
If NA, the maximum date in \code{hop} is used.}
}
\value{
A hop object.
}
\description{
Filters a hop object by date.
}
\examples{
\dontrun{
newhop <- hop_date_filter(myhop, NA, "2010-01-01")
}
}
\seealso{
Other hisafe helper functions: \code{\link{hisafe_params}},
  \code{\link{hisafe_profiles}}, \code{\link{hop_filter}},
  \code{\link{hop_merge}}, \code{\link{hop_rename}}
}
