% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hop_filter}
\alias{hop_filter}
\title{Filter a hop object by SimulationName}
\usage{
hop_filter(hop, simu.names)
}
\arguments{
\item{hop}{A object of class hop or face.}

\item{simu.names}{A character vector of the SimulationNames to keep. If "all", no filtering occurs.}
}
\value{
A hop object.
}
\description{
Filters a hop object by SimulationName
}
\examples{
\dontrun{
newhop <- hop_filter(myhop, c("Sim_1", "Sim_2"))
}
}
\seealso{
Other hisafe helper functions: \code{\link{hip_params}},
  \code{\link{hisafe_profiles}},
  \code{\link{hop_date_filter}}, \code{\link{hop_merge}},
  \code{\link{hop_params}}, \code{\link{hop_rename}}
}
