% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hop_merge}
\alias{hop_merge}
\title{Merge multiple hop objects}
\usage{
hop_merge(...)
}
\arguments{
\item{...}{Any number of individual hop objects to merge.}
}
\value{
Returns a hop object.
}
\description{
Merges multiple hop objects, renaming simulation names if there are duplicates
}
\examples{
\dontrun{
new_hop <- hop_merge(hop1, hop2, hop3)
}
}
\seealso{
Other hisafe helper functions: \code{\link{copy_hisafe_template}},
  \code{\link{hip_params}}, \code{\link{hisafe_info}},
  \code{\link{hisafe_profiles}}, \code{\link{hop_filter}},
  \code{\link{hop_params}}, \code{\link{hop_rename}},
  \code{\link{is_hip}}, \code{\link{is_hop}},
  \code{\link{join_profiles}}, \code{\link{profile_check}},
  \code{\link{variable_check}},
  \code{\link{which_profiles}}
}
\concept{hisafe helper functions}
