% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hop_params}
\alias{hop_params}
\title{Display Hi-sAFe output variables}
\usage{
hop_params(variables = "all", search = FALSE, quiet = FALSE)
}
\arguments{
\item{variables}{A character vector of specific Hi-sAFe output variables of which to display details.
Can also be a regular expression for which to search in Hi-sAFe output variable names.
If "all", the default, then a data.frame (tibble) containing all Hi-sAFe output variables is returned.}

\item{search}{Logical indicating whether \code{variables} should be treated as a regular expression and
searched for in the variable names rather than matched literally.}

\item{quiet}{Logical indicating whether or not to supress output printed to console.}
}
\value{
If \code{variables} is "all", the default, then a data.frame (tibble) containing all Hi-sAFe output variables
is returned. Otherwise, this data.frame is invisibly returned.
}
\description{
Displays Hi-sAFe output variables, their definitions, and thier units.
}
\examples{
\dontrun{
hop_params()                 # details of all variables
hop_params("carbonBranches") # details of cellWidth parameter
}
}
\seealso{
Other hisafe helper functions: 
\code{\link{copy_hisafe_template}()},
\code{\link{get_pheno_dates}()},
\code{\link{get_pruning_dates}()},
\code{\link{hip_params}()},
\code{\link{hisafe_info}()},
\code{\link{hisafe_profiles}()},
\code{\link{hop_filter}()},
\code{\link{hop_merge}()},
\code{\link{hop_rename}()},
\code{\link{is_face}()},
\code{\link{is_hip}()},
\code{\link{is_hop}()},
\code{\link{join_profiles}()},
\code{\link{make_rel_years}()},
\code{\link{profile_check}()},
\code{\link{variable_check}()},
\code{\link{which_profiles}()}
}
\concept{hisafe helper functions}
