% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_hip}
\alias{is_hip}
\title{Check if an object is of class hip}
\usage{
is_hip(hip, error = FALSE)
}
\arguments{
\item{x}{An object to check.}
}
\value{
A logical.
}
\description{
Checks if an object is of class hip
}
\examples{
\dontrun{
is_hip(myhip)
}
}
\seealso{
Other hisafe helper functions: \code{\link{hip_params}},
  \code{\link{hisafe_profiles}}, \code{\link{hop_filter}},
  \code{\link{hop_merge}}, \code{\link{hop_params}},
  \code{\link{hop_rename}}, \code{\link{is_hop}},
  \code{\link{profile_check}}, \code{\link{variable_check}}
}
