% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_hop}
\alias{is_hop}
\title{Check if an object is of class hop}
\usage{
is_hop(hop, error = FALSE)
}
\arguments{
\item{x}{An object to check.}
}
\value{
A logical.
}
\description{
Checks if an object is of class hop.
}
\examples{
\dontrun{
is_hop(myhop)
}
}
\seealso{
Other hisafe helper functions: \code{\link{copy_hisafe_template}},
  \code{\link{get_pheno_dates}},
  \code{\link{get_pruning_dates}},
  \code{\link{hip_params}}, \code{\link{hisafe_info}},
  \code{\link{hisafe_profiles}}, \code{\link{hop_filter}},
  \code{\link{hop_merge}}, \code{\link{hop_params}},
  \code{\link{hop_rename}}, \code{\link{is_hip}},
  \code{\link{join_profiles}}, \code{\link{profile_check}},
  \code{\link{variable_check}},
  \code{\link{warn_unequal_lengths}},
  \code{\link{which_profiles}}
}
\concept{hisafe helper functions}
