% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{join_profiles}
\alias{join_profiles}
\title{Join multiple hop profiles together}
\usage{
join_profiles(hop, profiles, ...)
}
\arguments{
\item{hop}{An object of class hop or face.}

\item{profiles}{The profiles to join}

\item{...}{Other arguments passed to \code{\link{hop_filter}}}
}
\value{
A tibble (data.frame)
}
\description{
Joins multiple hop profiles together into a single tibble (data.frame) using \code{dplyr::full_join()}.
When joining the 'cells' profile with the 'trees', 'plot', or 'climate' profiles, only numeric columns from the 'cells'
profile are kept, and values are averaged across all cells before joining.
When joining the 'voxels' profile with the 'cells' profile, only numeric columns from the 'voxels'
profile are kept, and values are summed across all voxels in each cell before joining.
The 'voxels' profile cannot be joinged with the 'trees', 'plot', or 'climate' profiles.
}
\examples{
\dontrun{
my.df <- join_profiles(hop, c("trees", "plot"))
}
}
\seealso{
Other hisafe helper functions: \code{\link{copy_hisafe_template}},
  \code{\link{hip_params}}, \code{\link{hisafe_info}},
  \code{\link{hisafe_profiles}}, \code{\link{hop_filter}},
  \code{\link{hop_merge}}, \code{\link{hop_params}},
  \code{\link{hop_rename}}, \code{\link{is_hip}},
  \code{\link{is_hop}}, \code{\link{profile_check}},
  \code{\link{variable_check}},
  \code{\link{which_profiles}}
}
\concept{hisafe helper functions}
