% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{layer_init_params}
\alias{layer_init_params}
\title{Generate soil layer initialization table for define_hisafe}
\usage{
layer_init_params(template, ...)
}
\arguments{
\item{template}{A character string of the path to the directory containing the template set of Hi-sAFe simulation folders/files to use.
See \code{\link{define_hisafe}} for more details.}

\item{...}{Any parameters of Hi-sAFe soil layer initialization table:
 \itemize{
 \item{"waterContent"}{}
 \item{"no3Concentration"}{}
 \item{"nh4concentration"}{}
}}
}
\value{
A list containing a Hi-sAFE soil layer initialization table.
}
\description{
Generates a soil layer initialization table suitable for passing to \code{\link{define_hisafe}}.
Any passed parameters modify the table in the provided template.
}
\examples{
\dontrun{
layer.init <- layer_init_params(template = "agroforestry", waterContent = 0.3)
}
}
\seealso{
Other hisafe definition functions: \code{\link{define_hisafe_file}},
  \code{\link{define_hisafe}}, \code{\link{layer_params}},
  \code{\link{root_init_params}},
  \code{\link{tree_init_params}}
}
