% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{layer_params}
\alias{layer_params}
\title{Generate soil layer table for define_hisafe}
\usage{
layer_params(template, ...)
}
\arguments{
\item{template}{A character string of the path to the directory containing the template set of Hi-sAFe simulation
folders/files to use.
See \code{\link{define_hisafe}} for more details.}

\item{...}{Any parameters of Hi-sAFe soil layer table:
 \itemize{
 \item{"thick"}{}
 \item{"sand"}{}
 \item{"clay"}{}
 \item{"limeStone"}{}
 \item{"organicMatter"}{}
 \item{"partSizeSand"}{}
 \item{"stone"}{}
 \item{"stoneType"}{}
 \item{"infiltrability"}{}
}}
}
\value{
A list containing a Hi-sAFE soil layer table.
}
\description{
Generates a soil layer table suitable for passing to \code{\link{define_hisafe}}.
The output of this function is always passed to the \code{\link{define_hisafe}} via the \code{layers} argument
(see example below). Any passed parameters modify the table in the provided template.
}
\examples{
\dontrun{
hip <- define_hisafe(path = getwd(),
                     layer.initialization = layer_init_params(template     = "agroforestry",
                                                              partSizeSand = 200))
}
}
\seealso{
Other hisafe definition functions: 
\code{\link{define_hisafe_file}()},
\code{\link{define_hisafe}()},
\code{\link{layer_init_params}()},
\code{\link{root_init_params}()},
\code{\link{tree_init_params}()}
}
\concept{hisafe definition functions}
