% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LER.R
\name{LER}
\alias{LER}
\title{Plot LER for yield, light, water, or nitrogen}
\usage{
LER(face, cycle = "yield", timescales = c("Annual", "Cumulative"),
  components = c("LER", "Trees", "Crop"), color.palette = c("black",
  "#E69F00", "#009E73"), plot = TRUE, ...)
}
\arguments{
\item{face}{An object of class "face".}

\item{cycle}{One of "yield", "nitrogen", "water", or "light".}

\item{timescales}{One or both of "Annual" and "Cumulative", indicating which timescales to plot.}

\item{components}{One or more of "LER", "Trees", and "Crop, indicating which components to plot.}

\item{color.palette}{A character string of hex values or R standard color names defining the color palette to use in plots with multiple simulations.}

\item{plot}{If \code{TRUE}, the default, a ggplot object is returned. If \code{FALSE}, the data that would create the plot is returned.}

\item{...}{Other arguments passed to \code{\link{plot_hisafe_cycle_annual}}.}
}
\value{
If \code{plot = TRUE}, returns a ggplot object, otherwise the data that would create the plot is returned.
}
\description{
Plot LER for yield, light, water, or nitrogen.
}
\examples{
\dontrun{
ler.plot <- LER(hop, "yield")
}
}
