% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LER.R
\name{LER_summary}
\alias{LER_summary}
\title{Plot LER for yield, light, water, or nitrogen}
\usage{
LER_summary(face, metrics = c("yield", "light", "water", "nitrogen"),
  fixed.scale = TRUE, nrow = 4, output.path = NULL,
  plot.labels = NULL, legend.inside = TRUE, ...)
}
\arguments{
\item{face}{An object of class "face".}

\item{metrics}{A character vector indicating which LER metrics to include.
Supported metrics include: 'yield', 'light', 'water', and 'nitrogen.}

\item{fixed.scale}{A logical indicating wether the y-axes of all panels should have have the same scale.}

\item{nrow}{A numeric value of the number of rows in which to arrange plots.}

\item{output.path}{A character string indicating the path to the directory where plots should be saved.
If \code{NULL}, the experiment/simulation path is read from the hop object, and a directory is created there called "analysis".
The plot wil be saved in this directory as "LER_summary.jpg".}

\item{plot.labels}{A character vector of labels to label upper-right corner of plot panels.
If \code{NULL}, no labels will be added.}

\item{legend.inside}{A logical indicating whether the legend should be placed inside the plots.}

\item{...}{Other arguments passed to \code{\link{LER}}.}
}
\value{
Invisibly retruns an egg object.
}
\description{
Plot LER for yield, light, water, or nitrogen.
}
\examples{
\dontrun{
ler.plot <- summary_LER(face, "./")
}
}
\seealso{
Other hisafe analysis functions: \code{\link{LER}},
  \code{\link{analyze_hisafe}},
  \code{\link{cycle_summary}}, \code{\link{hisafe_budget}},
  \code{\link{plot_hisafe_cycle_bar}},
  \code{\link{plot_hisafe_cycle_ts}},
  \code{\link{stics_budget_comp}}, \code{\link{write_hop}}
}
\concept{hisafe analysis functions}
