% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{modify_table}
\alias{modify_table}
\title{Build tables for table param functions}
\usage{
modify_table(args, supported.args, numeric.args, character.args, positive.args,
  perc.args, table.name, template)
}
\arguments{
\item{args}{From table param function}

\item{supported.args}{From table param function}

\item{numeric.args}{From table param function}

\item{character.args}{From table param function}

\item{positive.args}{From table param function}

\item{perc.args}{From table param function}

\item{table.name}{From table param function}

\item{template}{From table param function}
}
\value{
A data.frame (tibble) containing a Hi-sAFE soil layer table.
}
\description{
Builds tables for table param functions
}
