% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{plot_daily_light}
\alias{plot_daily_light}
\title{Plot daily timeseries of nitrogen, water, or light capture}
\usage{
plot_daily_light(hop, years, simu.names = "all", color.palette = NULL,
  plot = TRUE)
}
\arguments{
\item{hop}{An object of class hop or face.}

\item{years}{A numeric vector of the calendar years to include. If more than one year is provided, years are used as facets, and only a single value can be supplied to \code{simu.names}. If "all" is supplied, then all simulated years are used.}

\item{simu.names}{A character vector of the SimulationNames within \code{hop} to include. The default, "all", includes all simulations.
If more than one value is supplied to \code{years}, then a single value must be suppied to \code{simu.names}.}

\item{color.palette}{A character stirng of hex values or R standard color names defining the color palette to use in plots with multiple simulations.
If \code{NULL}, the default, then the default color palette is a color-blind-friendly color palette.}

\item{plot}{If \code{TRUE}, the default, a ggplot object is returned. If \code{FALSE}, the data that would create the plot is returned.}
}
\value{
If \code{plot = TRUE}, returns a ggplot object. If \code{plot = FALSE}, returns the data that would create the plot.
If \code{hop} contains more than one simulation, the plot will be faceted by SimulationName.
}
\description{
Plots a daily timeseries of water, nitrogen, or light, capture
}
\seealso{
Other hisafe analysis functions: \code{\link{plot_cycle}}
}
