% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hisafe_bg}
\alias{plot_hisafe_bg}
\title{Plot belowground tree root depth & water table}
\usage{
plot_hisafe_bg(hop, simu.names = "all", years = "all", tree.ids = 1,
  sen = c("fine", "coarse"), plot = TRUE)
}
\arguments{
\item{hop}{An object of class hop.}

\item{simu.names}{A character vector of the SimulationNames within \code{hop} to include. Use "all" to include all available values.}

\item{years}{A numeric vector of the years within \code{hop} to include. Use "all" to include all available values.}

\item{tree.ids}{A numeric vector indicating a subset of tree ids to plot. Only one tree can be plotted at a time.}

\item{sen}{A character vector indicated whether or not to plot the fine and coarse root senesence. Any value other than than "fine" and/or "coarse" will prevent them from being plotted.}

\item{plot}{If \code{TRUE}, the default, a ggplot object is returned. If \code{FALSE}, the data that would create the plot is returned.}
}
\value{
If \code{plot = TRUE}, returns a ggplot object, otherwise the data that would create the plot is returned.
}
\description{
Plots a daily timeseries of tree maximum rooting depth, water table depth, and root senescence by anoxia.
}
\examples{
\dontrun{
plot_hisafe_bg(hop)
}
}
\seealso{
Other hisafe plot functions: \code{\link{plot_hisafe_annualcells}},
  \code{\link{plot_hisafe_cells}},
  \code{\link{plot_hisafe_monthcells}},
  \code{\link{plot_hisafe_scene}},
  \code{\link{plot_hisafe_ts}},
  \code{\link{plot_hisafe_voxels}}
}
\concept{hisafe plot functions}
