% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hisafe_cells}
\alias{plot_hisafe_cells}
\title{Tile plot of Hi-sAFe cells output variable}
\usage{
plot_hisafe_cells(
  hop,
  variable,
  dates,
  start.date = NULL,
  rel.dates = dates,
  simu.names = "all",
  X = NA,
  plot.x = "x",
  N.arrow = "#000000",
  trees = TRUE,
  canopies = TRUE,
  plot = TRUE,
  mem.max = 0,
  for.anim = FALSE,
  cumu.scale = TRUE
)
}
\arguments{
\item{hop}{An object of class hop.}

\item{variable}{A character string of the name of the variable to color the tiles.}

\item{dates}{A character vector (in the format "YYYY-MM-DD") or a vector of class Date of the dates to include.}

\item{start.date}{If \code{NULL}, the default, then no cumulation of values occurs. Otherwise, a character string (in the format "YYYY-MM-DD") or an object of class Date indicating the date from which cumulation of \code{variable} should be calculated.}

\item{rel.dates}{A character vector (in the format "YYYY-MM-DD") or a vector of class Date of the dates from which to scale \code{variable}.
In the plot, \code{variable} will be scaled to be between the minimum and maximum values of \code{variable} across these dates.}

\item{simu.names}{A character string containing the SimulationNames to include. Use "all" to include all available values.}

\item{X}{A numeric vector of the x values of the cells to include. If \code{NA}, the default, then all x values are used.}

\item{plot.x}{Either "x" or "y", indicating which axis of the simulation scene should be plotted on the x-axis of the plot.}

\item{N.arrow}{A character string indicating the color of the north arrow. String must be a valid color name passed to ggplot2.
Use \code{NA} to not plot the north arrow.}

\item{trees}{Logical indicating if a point should be plotted at the location of each tree.}

\item{canopies}{Logical indicating if an elipsoid should be plotted representing the size of each tree canopy.}

\item{plot}{If \code{TRUE}, the default, a ggplot object is returned. If \code{FALSE}, the data that would create the plot is returned.}

\item{mem.max}{An integer specifying the maximum number of days into the past to search
for cell data when no data is available for a given date within \code{hop}.}

\item{for.anim}{If \code{TRUE}, the plot formatting is simplified for use in animations.}

\item{cumu.scale}{If \code{TRUE}, the color scale is determined by all dates from \code{start.date} to \code{dates}.
If \code{FALSE}, the color scale is determined only by \code{rel.dates}. Only applies when \code{start.dates} is not \code{NULL}.}
}
\value{
Returns a ggplot object.
}
\description{
Plots a tile plot of a single Hi-sAFe cells output variable.
SimulationName is used as the column facet. Date is used as the row facet.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe(path = "./")

# You can create a tile plot of relativeDirectParIncident:
tile.plot <- plot_hisafe_cells(mydata, "relativeDirectParIncident", paste(1998, 6:8, 1, sep = "-"))

# Once you have the plot object, you can display it and save it:
tile.plot
ggsave_fitmax("tiled_relativeDirectParIncident.png", tile.plot)
}
}
\seealso{
Other hisafe plot functions: 
\code{\link{plot_hisafe_annualcells}()},
\code{\link{plot_hisafe_bg}()},
\code{\link{plot_hisafe_monthcells}()},
\code{\link{plot_hisafe_scene}()},
\code{\link{plot_hisafe_tstress}()},
\code{\link{plot_hisafe_ts}()},
\code{\link{plot_hisafe_voxels}()}
}
\concept{hisafe plot functions}
