% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{plot_hisafe_cycle}
\alias{plot_hisafe_cycle}
\title{Plot annual barchart of carbon, nitrogen, water, or light cycle}
\usage{
plot_hisafe_cycle(hop, cycle, simu.names = "all", time.lim = NULL,
  color.palette = NULL, plot = TRUE)
}
\arguments{
\item{hop}{An object of class hop or face.}

\item{cycle}{One of "carbon", "nitrogen", "water", or "light".}

\item{simu.names}{A character vector of the SimulationNames within \code{hop} to include. Use "all" to include all available values.}

\item{time.lim}{A numeric vector of length two providing the \code{c(minimum, maximum)} of years (absolute or since planting) to plot.
If no input, the full available time range is plotted. Use \code{NA} to refer to the start or end of the simulation.}

\item{color.palette}{A character stirng of hex values or R standard color names defining the color palette to use in plots with multiple simulations.
If \code{NULL}, the default, then the default color palette is a color-blind-friendly color palette.}

\item{plot}{If \code{TRUE}, the default, a ggplot object is returned. If \code{FALSE}, the data that would create the plot is returned.}
}
\value{
If \code{plot = TRUE}, returns a ggplot object. If \code{plot = FALSE}, returns the data that would create the plot.
If \code{hop} contains more than one simulation, the plot will be faceted by SimulationName.
}
\description{
Plots an annual barchart of the water, nitrogen, light, or carbon cycle
}
\examples{
\dontrun{
# To plot the simulated water cycle:
water.plot <- plot_hisafe_cycle(myhop, "water")

# Once you have the plot object, you can display it and save it:
water.plot
ggplot2::ggsave("water_cycle.png", water.plot)
}
}
\seealso{
Other hisafe analysis functions: \code{\link{plot_hisafe_use}}
}
