% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{plot_hisafe_cycle_daily}
\alias{plot_hisafe_cycle_daily}
\title{Plot daily timeseries of major cycles}
\usage{
plot_hisafe_cycle_daily(hop, cycle, years, simu.names = "all",
  tree.ids = "all", doy.lim = c(1, 366), color.palette = NULL,
  plot = TRUE)
}
\arguments{
\item{hop}{An object of class hop or face.}

\item{cycle}{One of "carbon", "nitrogen", "water", or "light".}

\item{years}{A numeric vector of the calendar years to include.
If more than one year is provided, years are used as facets, and only a single value can be supplied to \code{simu.names}.
Use "all" to include all available values.}

\item{simu.names}{A character vector of the SimulationNames within \code{hop} to include. Use "all" to include all available values.
If more than one value is supplied to \code{years}, then a single value must be suppied to \code{simu.names}.}

\item{tree.ids}{A numeric vector indicating a subset of tree ids to plot. Use "all" to include all available values.
This only applies when \code{cycle} is "carbon".}

\item{doy.lim}{A numeric vector of length two providing the \code{c(minimum, maximum)} of julian days to plot.}

\item{color.palette}{A character stirng of hex values or R standard color names defining the color palette to use in plots with multiple simulations.
If \code{NULL}, the default, then the default color palette is a color-blind-friendly color palette.}

\item{plot}{If \code{TRUE}, the default, a ggplot object is returned. If \code{FALSE}, the data that would create the plot is returned.}
}
\value{
If \code{plot = TRUE}, returns a ggplot object. If \code{plot = FALSE}, returns the data that would create the plot.
If \code{hop} contains more than one simulation, the plot will be faceted by SimulationName.
}
\description{
Plots a daily timeseries of carbon pools, water extraction, nitrogen extraction, or light capture
}
\examples{
\dontrun{
# To plot the simulated light capture in the year 2000:
light.capture.plot <- plot_hisafe_use(myhop, "light", 2000)

# Once you have the plot object, you can display it and save it:
water.plot
ggsave_fitmax("light_capture.png", light.capture.plot)
}
}
\seealso{
Other hisafe analysis functions: \code{\link{analyze_hisafe}},
  \code{\link{plot_hisafe_cycle_annual}}
}
