% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hisafe_monthcells}
\alias{plot_hisafe_monthcells}
\title{Tile plot of Hi-sAFe monthCells output variable}
\usage{
plot_hisafe_monthcells(hop, variable = "monthRelativeTotalParIncident",
  colfacet = "SimulationName", rowfacet = "Year", simu.names = "all",
  years = seq(min(hop$monthCells$Year), max(hop$monthCells$Year), 5),
  months = 6, plot.x = "x", trees = TRUE, canopies = TRUE,
  tree.simus = FALSE, plot = TRUE)
}
\arguments{
\item{hop}{An object of class hop.}

\item{variable}{A character string of the name of the variable to color the tiles.}

\item{colfacet}{One of "Year", "Month", or "SimulationName", indicating which variable to use for column faceting.}

\item{rowfacet}{One of "Year", "Month", or "SimulationName", indicating which variable to use for row faceting.}

\item{simu.names}{A character string containing the SimulationNames to include. Use "all" to include all available values.}

\item{years}{A numeric vector containing the years to include. Use "all" to include all available values.}

\item{months}{A numeric vector containing the months to include. Use "all" to include all available values.}

\item{plot.x}{Either "x" or "y", indicating which axis of the simulation scene should be plotted on the x-axis of the plot.}

\item{trees}{Logical indicating if a point should be plotted at the location of each tree.}

\item{canopies}{Logical indicating if an elipsoid should be plotted representing the size of each tree canopy.}

\item{tree.simus}{Logical indicating whether only simulations containing trees should be plotted.}

\item{plot}{If \code{TRUE}, the default, a ggplot object is returned. If \code{FALSE}, the data that would create the plot is returned.}
}
\value{
Returns a ggplot object.
}
\description{
Plots a tile plot of a single Hi-sAFe monthCells output variable.
}
\details{
This function is very picky! You can only facet by two of the three manipulable variables: SimulationName, Year, Month.
You must ensure that the one varibale not used for faceting is fixed at a single value.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe(path = "./")

# You can create a tile plot of monthDirectParIncident:
tile.plot <- plot_hisafe_monthcells(mydata, "monthDirectParIncident")

# The default settings use data from June and facet by Year and Simulation Name.
# If you instead want to just look at Year 20 and facet by Month and SimulationName:
tile.plot2 <- plot_hisafe_monthcells(mydata, "monthDirectParIncident",
                                     rowfacet = "SimulationName",
                                     colfacet = "Month",
                                     simu.names = "all",
                                     years = 2000,
                                     months = 1:12)

# Once you have the plot object, you can display it and save it:
tile.plot2
ggsave_fitmax("monthDirectParIncident.png", tile.plot2)
}
}
\seealso{
Other hisafe plot functions: \code{\link{plot_hisafe_annualcells}},
  \code{\link{plot_hisafe_bg}},
  \code{\link{plot_hisafe_cells}},
  \code{\link{plot_hisafe_scene}},
  \code{\link{plot_hisafe_ts}},
  \code{\link{plot_hisafe_voxels}}
}
\concept{hisafe plot functions}
