% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_hisafe_ts}
\alias{plot_hisafe_ts}
\title{Plot timeseries of Hi-sAFe output variable}
\usage{
plot_hisafe_ts(data, variable, time.class = "Annual")
}
\arguments{
\item{data}{An object of class \code{hop} or \code{hop-group} containing output data from one or more Hi-sAFe simulations.}

\item{variable}{A character string of the name of the variable to plot.}

\item{time.class}{If 'Annual', an annual timeseries is created. If 'Daily', a daily timeseries is created.}
}
\value{
A ggplot object. If the data is of class \code{hop-group} and contains data from more than one
Hi-sAFe simulation, the plot will contain multiple lines, colored and labeled by SimName. If the data
contains two more tree ids, the plot will be faceted by tree id.
}
\description{
Plots a daily or annual timeseries of a single Hi-sAFe output variable.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe_output("MySimulation", "./")

# You can create an annual timeseries of carbonCoarseRoots:
annual.plot <- plot_hisafe_ts(mydata, "carbonCoarseRoots", "Annual")

# For a daily timeseries instead:
daily.plot <- plot_hisafe_ts(mydata, "carbonCoarseRoots", "Daily")

# Once you have the plot object, you can display it and save it:
annual.plot
ggsave("annual_carbonCoarseRoots.png", annual.plot)
}
}
