% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hisafe_tstress}
\alias{plot_hisafe_tstress}
\title{Plot crop temperature and temperature stress regions}
\usage{
plot_hisafe_tstress(hop, simu.name, type = "lue", years = "all",
  months = "all", cell.ids = "all", facet.year = TRUE,
  facet.month = TRUE, temin = 0, teopt = 10, teoptbis = 20,
  temax = 40, tminremp = 0, tmaxremp = 30,
  shade.color = c("grey50", "grey70", "white"))
}
\arguments{
\item{hop}{An object of class hop.}

\item{simu.name}{A character string of the SimulationName to plot. Only one simulation can be plotted.}

\item{years}{A numeric vector of the years within \code{hop} to include. Use "all" to include all available values.}

\item{months}{A numeric vector of the months within \code{hop} to include. Use "all" to include all available values.
Only applies if \code{type} is 'lue'. If \code{type} is 'reprod', then all days during which grain filling occurs are used.}

\item{cell.ids}{A numeric vector of the values of idCell within \code{hop} to include. Use "all" to include all available values.}

\item{facet.year}{A logical indicating whether the plot should be faceted by year. This helps with seeing finer level detail.}

\item{facet.month}{A logical indicating whether the plot should be faceted by month. Only applies if \code{type} is 'lue'.}

\item{temin}{Minimum threshold temperature for growth in biomass. Only applies if \code{type} is 'lue'.}

\item{teopt}{Optimum temperature for growth in biomass. Only applies if \code{type} is 'lue'.}

\item{teoptbis}{Optimum temperature for growth in biomass (if plateau between \code{teopt} and \code{teoptbis}). Only applies if \code{type} is 'lue'.}

\item{temax}{Maximum threshold temperature for growth in biomass. Only applies if \code{type} is 'lue'.}

\item{tminremp}{Minimum temperature for grain filling. Only applies if \code{type} is 'reprod'.}

\item{tmaxremp}{Maximum temperature for grain filling. Only applies if \code{type} is 'reprod'.}

\item{shade.color}{A character vector of length 3 indicating the shading colors of the various stress regions.
Strings must be a valid color name passed to ggplot2.}
}
\value{
A ggplot object.
}
\description{
Plots histograms of crop temperature and relevant temperature stress regions.
If \code{type} is 'lue', then \code{hop$cells$cropTemperature} (mean crop temperature) is shown in relation to tempStressLue (FTEMP in STICS) regions.
If \code{type} is 'reprod', then \code{hop$cells$cropMinTemperature} (min crop temperature) and \code{hop$cells$cropMaxTemperature} (max crop temperature)
are shown in relation to tempStressGrainFilling (FTEMPREMP in STICS) regions.
}
\examples{
\dontrun{
plot_hisafe_tempstress(hop, "Sim_1")
}
}
\seealso{
Other hisafe plot functions: \code{\link{plot_hisafe_annualcells}},
  \code{\link{plot_hisafe_bg}},
  \code{\link{plot_hisafe_cells}},
  \code{\link{plot_hisafe_monthcells}},
  \code{\link{plot_hisafe_scene}},
  \code{\link{plot_hisafe_ts}},
  \code{\link{plot_hisafe_voxels}}
}
\concept{hisafe plot functions}
