% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hisafe_voxels}
\alias{plot_hisafe_voxels}
\title{Tile plot of Hi-sAFe voxels output variable}
\usage{
plot_hisafe_voxels(hop, variable, dates, plot = TRUE)
}
\arguments{
\item{hop}{An object of class hop.}

\item{variable}{A character string of the name of the variable to color the tiles.}

\item{dates}{A character vector containing the dates (yyyy-mm-dd) to include.}

\item{plot}{If \code{TRUE}, the default, a ggplot object is returned. If \code{FALSE}, the data that would create the plot is returned.}
}
\value{
Returns ggplot object.
}
\description{
Plots a tile plot of a single Hi-sAFe voxels output variable.
If a single date is provided, SimulationName is used as the column facet. Otherwise, Date is used as the column facet.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe(path = "./")

# You can create a tile plot of waterAvailable:
tile.plot <- plot_hisafe_voxels(mydata, "waterAvailable", paste(1998, 6:8, 1, sep = "-"))

# Once you have the plot object, you can display it and save it:
tile.plot
ggplot2::ggsave("tiled_waterAvailable.png", tile.plot)
}
}
\seealso{
Other hisafe plot functions: \code{\link{plot_hisafe_cells}},
  \code{\link{plot_hisafe_monthcells}},
  \code{\link{plot_hisafe_scene}},
  \code{\link{plot_hisafe_ts}}
}
