% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{profile_check}
\alias{profile_check}
\title{Check for existiance of profiles in a hop object}
\usage{
profile_check(hop, cycle, required.profiles)
}
\arguments{
\item{hop}{An object of class hop or face.}

\item{cycle}{Cycle name to print in error output.}

\item{required.profiles}{A character vector of the names of the profiles to check for.}
}
\value{
Throws an error if profiles are not found.
}
\description{
Checks for existiance of profiles in a hop object
Used within \code{\link{get_water_fluxes}}, \code{\link{get_nitrogen_fluxes}}, \code{\link{get_light_fluxes}}, \code{\link{get_carbon_pools}}
}
