% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_hisafe}
\alias{read_hisafe}
\title{Read output from one or more Hi-sAFe simulations}
\usage{
read_hisafe(hip = NULL, path = NULL, simu.names = "all",
  profiles = "all", show.progress = TRUE, max.size = 3e+08,
  date.min = NA, date.max = NA)
}
\arguments{
\item{hip}{An object of class "hip". To create a hip object see \code{\link{define_hisafe}}.
If \code{hip} is not provided, then \code{path} is required and the input data for the experiment is read from the experiment
summary .csv file created when building the experiment.}

\item{path}{A character string of the path to the directory containing the Hi-sAFe simulation folders.
If \code{hip} is not provided, then \code{path} is required. If \code{hip} is provided, \code{path} is ignored.}

\item{simu.names}{Names of the simulations to read. If "all", the default, then all simulations in the folder are read.}

\item{profiles}{A character vector of the names of Hi-sAFe output profiles to read.
#' If "all" the default, reads all supported Hi-sAFe output profiles. For currently supported profiles see: \code{\link{hisafe_profiles}}}

\item{show.progress}{Logical indicating whether progress messsages should be printed to the console.}

\item{max.size}{The maximum file size (in bytes) that should be read. Files larger than this value will be ignored, with a warning.}

\item{date.min}{A character string of the minimum date to keep, in the format "YYYY-MM-DD".
If NA, the minimum date in the output data is used.}

\item{date.max}{A character string of the maximum date to keep, in the format "YYYY-MM-DD".
If NA, the maximum date in the output data is used.}
}
\value{
An object of class "hop". This is a list of 16 data frames (tibbles):
\itemize{
 \item{annualtree}
 \item{annualcrop}
 \item{annualplot}
 \item{trees}
 \item{plot}
 \item{climate}
 \item{monthCells}
 \item{cells}
 \item{roots}
 \item{voxels}
 \item{variables}{ - variable descriptions and units from all read profiles}
 \item{plot.info}{ - plot geometry data for each simulation}
 \item{tree.info}{ - tree species and location data for each simulation}
 \item{exp.plan}{ - the exp.plan of the hip object that generated the simulation}
 \item{path}{ - the paths to the simulation folders}
 \item{exp.path}{ - the path to the experiment folder}
}
}
\description{
Reads the designated output profiles from one or more Hi-sAFe simulations
}
\examples{
\dontrun{
# Reading in Hi-sAFe simulation:
myexp <- read_hisafe(myhip)

# If only the annual tree data is required:
mytreeexp <- read_hisafe(myhip, profiles = "annualtree")
}
}
