% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_hisafe_example}
\alias{read_hisafe_example}
\title{Read example Hi-sAFe experiment output}
\usage{
read_hisafe_example(simu.names = c("monocrop", "agroforestry",
  "forestry"), profiles = c("plot", "plotDetail", "trees", "treesDetail",
  "cells", "cellsDetail", "voxelsMonth", "climate", "monthCells",
  "annualCells"), ...)
}
\arguments{
\item{simu.names}{A character vector of the names of the simulations to read. If "all", the default, then all simulations in the folder are read.}

\item{profiles}{A character vector of the names of Hi-sAFe output profiles to read.}

\item{...}{Other arguments passed to \code{\link{read_hisafe}}.}
}
\value{
An object of class "hop".
}
\description{
Reads in an example Hi-sAFe experiment. For more details see \code{\link{read_hisafe}}.
}
