% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{read_hisafe_group}
\alias{read_hisafe_group}
\title{Read output from a group of Hi-sAFe simulations}
\usage{
read_hisafe_group(exp.plan, folder, profiles = c("annualtree", "annualplot",
  "trees", "plot", "monthCells"))
}
\arguments{
\item{exp.plan}{A data frame containing the experimental plan used to generate the Hi-sAFe simulations. To create an experimental plan, see \code{\link{define_exp}}.}

\item{folder}{A character string of the path to the directory containing the Hi-sAFe simulation folders (which each contain the standard subdirectories with the outputs)}

\item{profiles}{A character vector of the names of Hi-sAFe output profiles to read. Defaults to reading the core Hi-sAFe output profiles.}
}
\value{
An object of class \code{hisafe-group}. This is a list of five data frames:
\code{annual}, \code{daily}, \code{spatial} (monthly cell data), \code{variables} (variable descriptions and units),
and \code{exp.plan} (the provided experimental plan). Data frames will be empty for any data classes not included in the specified profiles
}
\description{
Reads the designated output profiles from a group of Hi-sAFe simulations (i.e. an experiment).
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
myexp <- read_hisafe_group(MyExpPlan, "./")

# If only the annual tree data is required:
mytreeexp <- read_hisafe(MyExpPlan, "./", profiles = "annualtree")
}
}
