% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{read_hisafe_output_file}
\alias{read_hisafe_output_file}
\title{Read a single Hi-sAFe output profile}
\usage{
read_hisafe_output_file(profile, read.data = TRUE)
}
\arguments{
\item{profile}{A character string of the path to the profile to be read.}

\item{read.data}{If TRUE, data and variable descriptions are read. If FALSE, only variable descriptions are read.}
}
\value{
An list of two data frames: \code{data} contains the data from the profile; \code{variables} contains the variable descriptions.
}
\description{
Reads the designated output profiles from a single Hi-sAFe simulation.
}
