% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{read_param_file}
\alias{read_param_file}
\title{Reads parameter files}
\usage{
read_param_file(path)
}
\arguments{
\item{path}{A character string of the path to the file to read.}
}
\value{
A list containing all parameter values and constraints.
}
\description{
Reads parameter files for editing
Used within \code{\link{build_structure}}.
}
\seealso{
Other hisafe param functions: \code{\link{edit_param_element}},
  \code{\link{edit_param_file}},
  \code{\link{write_param_file}}
}
\keyword{internal}
