% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_simulation}
\alias{read_simulation}
\title{Read output from a single Hi-sAFe simulation}
\usage{
read_simulation(simu.name, hip, path, profiles, show.progress, read.inputs,
  max.size)
}
\arguments{
\item{simu.name}{The \code{SimulationName} of the Hi-sAFe simulation to read. This must be the same as the name of the Hi-sAFe simulation folder.
Cannot provided both \code{hip} and \code{simu.name}.}

\item{hip}{An object of class "hip". To create a hip object see \code{\link{define_hisafe}}.
Cannot provided both \code{hip} and \code{simu.name}.
If \code{hip} is not provided, then \code{path} is required and the input data for the experiment is read from the experiment
summary .csv file created when building the experiment.}

\item{path}{A character string of the path to the directory containing the Hi-sAFe simulation folder.
If \code{hip} is not provided, then \code{path} is required. If both \code{hip} and \code{path} are provided, \code{path} is used.}

\item{profiles}{A character vector of the names of Hi-sAFe output profiles to read.
If "all" the default, reads all supported Hi-sAFe output profiles. For currently supported profiles see: \code{\link{hisafe_profiles}}}

\item{show.progress}{Logical indicating whether progress messsages should be printed to the console.}

\item{read.inputs}{Logical indicating whether data should be read from the .PLD and .SIM files for the tree.info and plot.info data slots.
Setting this to \code{FALSE} severly limits hisafer's plotting/analysis capacity.
Should only be set to \code{FALSE} if input files are corrupted.}

\item{max.size}{The maximum file size (in bytes) that should be read. Files larger than this value will be ignored, with a warning.}
}
\value{
An object of class "hop". This is a list of 15 data frames (tibbles):
\itemize{
 \item{annualTrees}
 \item{annualCells}
 \item{annualPlot}
 \item{trees}
 \item{plot}
 \item{climate}
 \item{monthCells}
 \item{cells}
 \item{voxels}
 \item{variables}{ - variable descriptions and units from all profiles}
 \item{plot.info}{ - plot geometry data}
 \item{tree.info}{ - tree species and location data}
 \item{exp.plan}{ - the exp.plan of the hip object that generated the simulation}
 \item{path}{ - the path to the simulation folder}
}
}
\description{
Reads the designated output profiles from a single Hi-sAFe simulation. Called from within \code{\link{read_hisafe}}.
}
\keyword{internal}
