% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_table_hisafe}
\alias{read_table_hisafe}
\title{Read from a Hi-sAFe output profile}
\usage{
read_table_hisafe(file, ...)
}
\arguments{
\item{file}{A character string of the path to the file to be read.}

\item{...}{Any other arguements passed to \code{read.table}}
}
\value{
A data frame.
}
\description{
Customized read.table call to read from a Hi-sAFe output profile.
}
\keyword{internal}
