% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_tree_info}
\alias{read_tree_info}
\title{Read tree information from a Hi-sAFe pld file}
\usage{
read_tree_info(path, simu.name)
}
\arguments{
\item{path}{A character string of the path to the directory containing the simulation folder.}

\item{simu.name}{A character string of the simualation name.}
}
\value{
A data frame (tibble) containing tree id, species, x and y.
}
\description{
Reads tree information from a Hi-sAFe pld file. Used by \code{\link{read_hisafe}}.
}
