% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{read_weather}
\alias{read_weather}
\title{Read weather data from a Hi-sAFe .WTH file.}
\usage{
read_weather(path)
}
\arguments{
\item{path}{A character string of the path (relative or absolute) to the .WTH file.}
}
\value{
A data frame (tibble) of the weather data.
}
\description{
Reads weather data from a Hi-sAFe .WTH file.
}
\examples{
\dontrun{
wth <- read_weather("./my_site.wth")
}
}
\seealso{
Other hisafe weather functions: \code{\link{write_weather}}
}
