% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{root_init_params}
\alias{root_init_params}
\title{Generate root initialization table for define_hisafe}
\usage{
root_init_params(reps = 1, shape = 1, repartition = 3,
  paramShape1 = 0.6, paramShape2 = 0, paramShape3 = 0, amount = 0.5)
}
\arguments{
\item{reps}{Number of times to repeats the rows of the defined table.}

\item{shape}{Parameter of Hi-sAFe root initialization table.}

\item{repartition}{Parameter of Hi-sAFe root initialization table.}

\item{paramShape1}{Parameter of Hi-sAFe root initialization table.}

\item{paramShape2}{Parameter of Hi-sAFe root initialization table.}

\item{paramShape3}{Parameter of Hi-sAFe root initialization table.}

\item{amount}{Parameter of Hi-sAFe root initialization table.}
}
\value{
A list containing one or more data frames (tibbles).
}
\description{
Generates a root initialization table suitable for passing to \code{\link{define_hisafe}}.
}
\examples{
\dontrun{
root.init <- root_init_params()
}
}
\seealso{
Other hisafe definition functions: \code{\link{define_hisafe_file}},
  \code{\link{define_hisafe}},
  \code{\link{layer_init_params}},
  \code{\link{layer_params}},
  \code{\link{tree_init_params}}
}
