% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_hisafe}
\alias{run_hisafe}
\title{Run a Hi-sAFe simulation}
\usage{
run_hisafe(hip = NULL, path = NULL, simu.name = NULL,
  capsis.path = "/Applications/Capsis/")
}
\arguments{
\item{hip}{An object of class "hip" with a single simulation (row). Required if \code{path} and \code{simu.name} are not provided.}

\item{path}{A character string of the path to the folder containing the simulation folder. Required if \code{hip} is not provided.}

\item{simu.name}{Name of the simulation to run. Required if \code{hip} is not provided.}

\item{capsis.path}{A character string of the path to the Capsis folder}
}
\value{
Invisibly returns a console log from simulation run.
}
\description{
Runs a Hi-sAFe simulation - calls Hi-sAFe from command line.
}
\examples{
\dontrun{
# Once a group Hi-sAFe simulations (experiment) is defined...
myexp <- define_hisafe(path = "./simulations")

# ...and built...
build_hisafe(myexp)

# the simulations can all be run:
run_hisafe(myexp)
}
}
\seealso{
Other hisafe run functions: \code{\link{run_hisafe_exp}}
}
