% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{shift_scene}
\alias{shift_scene}
\title{Flip scene about cell row}
\usage{
shift_scene(hop, profile, shift)
}
\arguments{
\item{hop}{An object of class hop.}

\item{profile}{Character string of the name of a hop profile.}

\item{shift}{The number of cell rows to shift the scene to the left.}
}
\description{
Flips scene about cell row for \code{\link{hisafe_slice}}
}
\keyword{internal}
