% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{shift_year}
\alias{shift_year}
\title{Shift Year column based on a provided JulianDay}
\usage{
shift_year(hop, doy.start)
}
\arguments{
\item{hop}{An object of class hop or a hop-like object.}

\item{doy.start}{The JulianDay [1-365] on which to start the annual cycle accounting. Use 'sim' to specify the starting JulianDay of the simulation.}
}
\value{
A hop object.
}
\description{
Shifts Year column based on a provided JulianDay.
}
\keyword{internal}
