% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\name{stics_budget_comp}
\alias{stics_budget_comp}
\title{Compare Hi-sAFe and STICS budget calcualations}
\usage{
stics_budget_comp(hop, cycle, simu.names = "all", years = "all",
  save.table = TRUE, output.path = NULL)
}
\arguments{
\item{hop}{An object of class hop or face.}

\item{cycle}{A character vector of the cycle to analyze Supported cycles include: 'water', and 'nitrogen.}

\item{simu.names}{A character vector of the SimulationNames within \code{hop} to include. Use "all" to include all available values.}

\item{years}{A numeric vector of the years within \code{hop} to include. Use "all" to include all available values.}

\item{save.table}{Logical indicating whether a table of comparison data should be saved to \code{output.path}.}

\item{output.path}{A character string indicating the path to the directory where plots should be saved.
If \code{NULL}, the experiment/simulation path is read from the hop object, and a directory is created there called "analysis/budgets".
The tables and plots will be saved in this directory.}
}
\value{
A data.frame (tibble) containing the budget comparison.
}
\description{
Compares Hi-sAFe and STICS budget calcualations for debugging purposes.
}
\examples{
\dontrun{
water.budget.comp <- stics_budget_comp(myhop, "water")
}
}
\seealso{
Other hisafe analysis functions: \code{\link{LER_summary}},
  \code{\link{LER}}, \code{\link{analyze_hisafe}},
  \code{\link{cycle_summary}}, \code{\link{hisafe_budget}},
  \code{\link{plot_hisafe_cycle_bar}},
  \code{\link{plot_hisafe_cycle_ts}},
  \code{\link{write_hop}}
}
\concept{hisafe analysis functions}
