% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_hisafe_monthcells}
\alias{theme_hisafe_monthcells}
\title{A \code{ggplot2} theme for Hi-sAFe monthCells tile plots}
\usage{
theme_hisafe_monthcells(base_size = 18)
}
\arguments{
\item{base_size}{A number indicating the base \code{ggplot} text size.}
}
\value{
A ggplot modifier object containing the theme.
}
\description{
The default ggplot2 theme for Hi-sAFe monthCells tile plots.
}
