% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_hisafe_ts}
\alias{theme_hisafe_ts}
\title{A \code{ggplot2} theme for Hi-sAFe timeseries plots}
\usage{
theme_hisafe_ts(base_size = 18)
}
\arguments{
\item{base_size}{A number indicating the base \code{ggplot} text size.}
}
\value{
A ggplot modifier object containing the theme.
}
\description{
The default ggplot2 theme for Hi-sAFe timeseries plots.
}
