% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{tree_init_params}
\alias{tree_init_params}
\title{Generate tree initialization table for define_hisafe}
\usage{
tree_init_params(species = "walnut-hybrid", age = 1, height = 1,
  crownBaseHeight = 0.5, truncatureRatio = 0, leafToFineRootsRatio = 0.5,
  crownRadius = 0.25, treeX = 0, treeY = 0)
}
\arguments{
\item{species}{Parameter of Hi-sAFe tree initialization table.}

\item{age}{Parameter of Hi-sAFe tree initialization table.}

\item{height}{Parameter of Hi-sAFe tree initialization table.}

\item{crownBaseHeight}{Parameter of Hi-sAFe tree initialization table.}

\item{truncatureRatio}{Parameter of Hi-sAFe tree initialization table.}

\item{leafToFineRootsRatio}{Parameter of Hi-sAFe tree initialization table.}

\item{crownRadius}{Parameter of Hi-sAFe tree initialization table.}

\item{treeX}{Parameter of Hi-sAFe tree initialization table.}

\item{treeY}{Parameter of Hi-sAFe tree initialization table.}
}
\value{
A list containing one or more data frames (tibbles).
}
\description{
Generates a tree initialization table suitable for passing to \code{\link{define_hisafe}}.
}
\examples{
\dontrun{
tree.init <- tree_init_params()
}
}
\seealso{
Other hisafe definition functions: \code{\link{define_hisafe_file}},
  \code{\link{define_hisafe}},
  \code{\link{layer_init_params}},
  \code{\link{layer_params}},
  \code{\link{root_init_params}}
}
