% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{variable_check}
\alias{variable_check}
\title{Check for existiance of variables in a hop object}
\usage{
variable_check(hop, profile, variables, error = FALSE)
}
\arguments{
\item{hop}{An object of class hop or face.}

\item{profile}{A character string of the name of the profile to check within.}

\item{variables}{A character vector of the names of the variables to check for.}

\item{error}{Logical indicating whehter or not an error should be thrown if any variables in \code{variables} are not found.}
}
\value{
A logical vector the same length as \code{variables} indicating whether each variable is found in \code{hop}.
If \code{error} is \code{TRUE}, this vector is returned invisibly.
}
\description{
Checks for existiance of variables within a profile of a hop object
}
\examples{
\dontrun{
variable_check(myhop, "trees", "carbonBranches")
}
}
\seealso{
Other hisafe helper functions: \code{\link{copy_hisafe_template}},
  \code{\link{get_pheno_dates}},
  \code{\link{get_pruning_dates}},
  \code{\link{hip_params}}, \code{\link{hisafe_info}},
  \code{\link{hisafe_profiles}}, \code{\link{hop_filter}},
  \code{\link{hop_merge}}, \code{\link{hop_params}},
  \code{\link{hop_rename}}, \code{\link{is_face}},
  \code{\link{is_hip}}, \code{\link{is_hop}},
  \code{\link{join_profiles}},
  \code{\link{make_rel_years}},
  \code{\link{profile_check}}, \code{\link{which_profiles}}
}
\concept{hisafe helper functions}
