% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{variety_params}
\alias{variety_params}
\title{Generate variety table for define_hisafe}
\usage{
variety_params(template, ...)
}
\arguments{
\item{template}{A character string of the path to the directory containing the template set of Hi-sAFe simulation
folders/files to use.
See \code{\link{define_hisafe}} for more details.}

\item{...}{Any parameters of Hi-sAFe variety table:
 \itemize{
 \item{"ID"}{}
 \item{"codevar"}{}
 \item{"stlevamf"}{}
 \item{"stamflax"}{}
 \item{"stlevdrp"}{}
 \item{"stflodrp"}{}
 \item{"stdrpdes"}{}
 \item{"pgrainmaxi"}{}
 \item{"adens"}{}
 \item{"croirac"}{}
 \item{"durvieF"}{}
 \item{"jvc"}{}
 \item{"sensiphot"}{}
 \item{"stlaxsen"}{}
 \item{"stsenlan"}{}
 \item{"nbgrmax"}{}
 \item{"stdrpmat"}{}
 \item{"afruitpot"}{}
 \item{"dureefruit"}{}
}}
}
\value{
A list containing a Hi-sAFE variety table.
}
\description{
Generates a variety table suitable for passing to \code{\link{define_hisafe}}.
The output of this function is always passed to \code{\link{define_hisafe}} via the \code{varieties} argument
(see example below). Any passed parameters modify the table in the provided template.
}
\examples{
\dontrun{
hip <- define_hisafe(path = getwd(),
                     varieties = variety_params(template = "monocrop",
                                                croirac  = 0.1))
}
}
\seealso{
Other hisafe definition functions: 
\code{\link{define_hisafe_file}()},
\code{\link{define_hisafe}()},
\code{\link{layer_init_params}()},
\code{\link{layer_params}()},
\code{\link{root_init_params}()},
\code{\link{tree_init_params}()}
}
\concept{hisafe definition functions}
