% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{vars_to_diag}
\alias{vars_to_diag}
\title{Determine diagnostic variables to plot}
\usage{
vars_to_diag(hop, profile, last.col)
}
\arguments{
\item{hop}{An object of class hop or face}

\item{profile}{A character string of the profile}

\item{last.call}{A character string of the name of the variable which is the last "header" variable added by Hi-sAFe}
}
\value{
A character vector of the names of variables to plot
}
\description{
Determines diagnostic variables to plot
}
\keyword{internal}
