% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visual_legend}
\alias{visual_legend}
\title{Create a legend for hisafe_visual()}
\usage{
visual_legend(hop, vars = list(crown.alpha = "leafArea", trunk.alpha =
  "carbonLabile", crop.alpha = "lai", yield.alpha = "eai", voxel.alpha =
  "totalTreeRootDensity", voxel.border = "cropRootDensity", voxel.L.size =
  "waterStock", voxel.C.size = "totalTreeCoarseRootBiomass", voxel.R.size =
  "mineralNitrogenStock", voxel.L.alpha = "totalTreeWaterUptake",
  voxel.C.alpha = "fineRootCost", voxel.R.alpha =
  "totalTreeNitrogenUptake"), cells.var = "relativeTotalParIncident")
}
\arguments{
\item{hop}{An object of class hop.}

\item{vars}{A list of variable names. See \code{\link{hisafe_slice}} for details.}

\item{cells.var}{A character string of the name of the variable ploted by \code{\link{plot_hisafe_cells}}.}
}
\value{
A ggplot object containing the legend.
}
\description{
Creates a legend for hisafe_visual().
}
