% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{warn_unequal_lengths}
\alias{warn_unequal_lengths}
\title{Provide warning if simulation lengths are unequal}
\usage{
warn_unequal_lengths(hop)
}
\arguments{
\item{hop}{An object of class hop or a hop-like object.}
}
\value{
A tibble (data.frame) containing the SimulationName and durations in years.
}
\description{
Provides warning if simulation lengths are unequal.
}
\seealso{
Other hisafe helper functions: \code{\link{copy_hisafe_template}},
  \code{\link{get_pheno_dates}},
  \code{\link{get_pruning_dates}},
  \code{\link{hip_params}}, \code{\link{hisafe_info}},
  \code{\link{hisafe_profiles}}, \code{\link{hop_filter}},
  \code{\link{hop_merge}}, \code{\link{hop_params}},
  \code{\link{hop_rename}}, \code{\link{is_hip}},
  \code{\link{is_hop}}, \code{\link{join_profiles}},
  \code{\link{profile_check}},
  \code{\link{variable_check}},
  \code{\link{which_profiles}}
}
\concept{hisafe helper functions}
\keyword{internal}
