% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{which_profiles}
\alias{which_profiles}
\title{Get which profiles exist in a hop}
\usage{
which_profiles(hop, profiles = NULL)
}
\arguments{
\item{hop}{An object of class hop or face.}

\item{profiles}{A character vector of the names of the profiles to check for.
If \code{NULL}, returns the names of all available profiles.}
}
\value{
A character vector of the available profiles in the hop.
}
\description{
Gets names of which profiles exist in a hop.
}
\examples{
\dontrun{
what_profiles(myhop, c("cells", "voxels"))
}
}
\seealso{
Other hisafe helper functions: 
\code{\link{copy_hisafe_template}()},
\code{\link{get_pheno_dates}()},
\code{\link{get_pruning_dates}()},
\code{\link{hip_params}()},
\code{\link{hisafe_info}()},
\code{\link{hisafe_profiles}()},
\code{\link{hop_filter}()},
\code{\link{hop_merge}()},
\code{\link{hop_params}()},
\code{\link{hop_rename}()},
\code{\link{is_face}()},
\code{\link{is_hip}()},
\code{\link{is_hop}()},
\code{\link{join_profiles}()},
\code{\link{make_rel_years}()},
\code{\link{profile_check}()},
\code{\link{variable_check}()}
}
\concept{hisafe helper functions}
