% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{write_hop}
\alias{write_hop}
\title{Write hop profiles to CSV files}
\usage{
write_hop(hop, profiles = "all", output.path = NULL)
}
\arguments{
\item{hop}{An object of class hop or face.}

\item{profiles}{The profiles which to each write as a CSV.}

\item{output.path}{A character string indicating the path to the directory where CSV files should be saved.
Plots aresaved in a subdirectory within this directory named \code{/analysis/combined_outputs/}.
If \code{NULL}, the experiment/simulation path is read from the hop object.}
}
\value{
Invisibly returns \code{hop}.
}
\description{
Writes hop profiles to CSV files.
}
\examples{
\dontrun{
write_hop(myhop)
}
}
\seealso{
Other hisafe analysis functions: 
\code{\link{LER_summary}()},
\code{\link{LER}()},
\code{\link{analyze_hisafe}()},
\code{\link{cycle_summary}()},
\code{\link{hisafe_budget}()},
\code{\link{plot_hisafe_cycle_bar}()},
\code{\link{plot_hisafe_cycle_ts}()},
\code{\link{stics_budget_comp}()}
}
\concept{hisafe analysis functions}
