% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{write_param_file}
\alias{write_param_file}
\title{Writes parameter files}
\usage{
write_param_file(param.list, path)
}
\arguments{
\item{path}{A character string of the path to the file to read.}

\item{.list}{A list containing all parameter values and constraints.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Writes parameter files after editing
Used within \code{\link{build_structure}}.
}
\seealso{
Other hisafe param functions: \code{\link{edit_param_element}},
  \code{\link{edit_param_file}},
  \code{\link{read_param_file}}
}
\keyword{internal}
