% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{write_weather}
\alias{write_weather}
\title{Write weather data to a Hi-sAFe .WTH file.}
\usage{
write_weather(data, path)
}
\arguments{
\item{data}{A data frame containing the weather data.}

\item{path}{A character string of the path (relative or absolute) to the .WTH file.}
}
\description{
Writes weather data to a Hi-sAFe .WTH file.
}
\examples{
\dontrun{
write_weather(wth, "./my_site_modified.wth")
}
}
\seealso{
Other hisafe weather functions: \code{\link{read_weather}}
}
