% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{analyze_hisafe}
\alias{analyze_hisafe}
\title{Shortcut to Hi-sAFe analysis}
\usage{
analyze_hisafe(hop, carbon = TRUE, light = TRUE, nitrogen = TRUE,
  water = TRUE, carbon.daily = TRUE, light.daily = TRUE,
  nitrogen.daily = TRUE, water.daily = TRUE, carbon.increment = TRUE)
}
\arguments{
\item{hop}{An object of class hop or face.}

\item{carbon}{Logical indicating if annual carbon plot should be made.}

\item{light}{Logical indicating if annual light plot should be made.}

\item{nitrogen}{Logical indicating if annual nitrogen plot should be made.}

\item{water}{Logical indicating if annual water plot should be made.}

\item{carbon.daily}{Logical indicating if daily carbon plots should be made.}

\item{light.daily}{Logical indicating if daily light plots should be made.}

\item{nitrogen.daily}{Logical indicating if daily nitrogens plot should be made.}

\item{water.daily}{Logical indicating if daily water plots should be made.}

\item{carbon.increment}{Logical indicating if daily carbon increment plots should be made.}
}
\value{
Invisibly returns \code{hop}.
}
\description{
Runs the various Hi-sAFe analysis functions from a single call.
}
\examples{
\dontrun{
analyze_hisafe(myhop)
}
}
\seealso{
Other hisafe analysis functions: \code{\link{LER_summary}},
  \code{\link{LER}}, \code{\link{cycle_summary}},
  \code{\link{plot_hisafe_cycle_annual}},
  \code{\link{plot_hisafe_cycle_daily}},
  \code{\link{write_hop}}
}
\concept{hisafe analysis functions}
