% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{build_cluster_script}
\alias{build_cluster_script}
\title{Builds a cluster shell script}
\usage{
build_cluster_script(hip = NULL, simu.names = NULL,
  script.path = NULL, script.name = NULL, simu.prefix = NULL,
  model.version = "Capsis4", launch.call = "ScriptGen",
  default.folder = "", cluster.path, email.type = "END",
  email = NULL, num.cores = NULL, max.per.node = NULL,
  mem.spec = NULL)
}
\arguments{
\item{hip}{A hip object. If \code{NULL}, then \code{simu.names} and \code{script.path} are required.}

\item{simu.names}{A character vector of the names of the simulations to create the script for.
This must match the names of .SIM files as well as the names of the folders that files are in.}

\item{script.path}{A character string of the local path for where to create the cluster script file.
If \code{NULL}, \code{hip} must be provided, and the path used will be the path element of the hip object.}

\item{script.name}{A character string of the name of the cluster script file to call.}

\item{simu.prefix}{A character string of the SimulationName prefix.}

\item{model.version}{A character string of the name of the Capsis folder (version) to call on the cluster.}

\item{launch.call}{The name of the safe lanuch script to use (one of 'ScriptGen' or 'ScriptCalib')}

\item{default.folder}{The folder in safe/data/SimSettings to use for parameter files that are not found in the simulation folder.}

\item{cluster.path}{A character string of the path on the cluster where the simulation folder is located.}

\item{email.type}{Type of email notification. Valid type values are NONE, BEGIN, END, FAIL, REQUEUE, ALL.
Multiple values may be specified in a comma separated list within a single character string (e.g. "BEGIN,END").}

\item{email}{A character string of the email address to notify with cluster job updates. Use \code{NULL} for no email generation.}

\item{num.cores}{The number of cores to request from the cluster. Use \code{NULL} to make no specification.}

\item{max.per.node}{The maximum numver of tasks allowed per node on the cluster. This allows memory to be allocated to each task.
Use \code{NULL} to make no specification.}

\item{mem.spec}{An integer indicating the maximum memory use (Mb) permitted for each individual instance of Capsis/Hi-sAFe.
Use \code{NULL} to make no specification.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Builds a shell script to initialize Hi-sAFe simulations on the MUSE cluster.
}
\examples{
\dontrun{
mysim <- build_cluster_script("mysim",
                              script.path  = "/simulations/",
                              cluster.path = "/nfs/work/hisafe/simulations/user/",
                              email        = "me@work.com")
}
}
\concept{hisafe cluster functions}
