% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_hisafe}
\alias{build_hisafe}
\title{Builds a Hi-sAFe simulation}
\usage{
build_hisafe(hip, path, export.profiles = "all", saveProjectOption = FALSE,
  part.of.exp = FALSE)
}
\arguments{
\item{hip}{A single simulation (row) from an object of class hip.}

\item{path}{A character string of the path to the simulation folder.}

\item{export.profiles}{A character vector of output profiles the simulation to export.}

\item{saveProjectOption}{Logical, sets the saveProjectOption parameter in the .sim file.}

\item{part.of.exp}{Logical, indicates if \code{build_hisafe} was called from within \code{build_hisafe_exp}. Do not change this!}
}
\value{
Invisibly returns a list containing the original hip object and supplied path.
}
\description{
Builds a Hi-sAFe simulation - creates the folder structure and input files.
}
\examples{
\dontrun{
# Once a single Hi-sAFe simulation is defined:
myexp <- define_hisafe()

# Building the experiment folder structure & files:
build_hisafe(myexp, "./simulations")
}
}
\seealso{
Other hisafe build functions: \code{\link{build_hisafe_exp}}
}
