% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_hisafe}
\alias{build_hisafe}
\title{Builds a Hi-sAFe simulation or experiment}
\usage{
build_hisafe(hip, plot.scene = TRUE, summary.files = TRUE)
}
\arguments{
\item{hip}{An object of class "hip". To create a hip object see \code{\link{define_hisafe}}.}

\item{plot.scene}{Logical indicating whether \code{\link{plot_hisafe_scene}} should be used to export plots of each scene during the build.}

\item{summary.files}{Logical indicating whether or not to write out summary .CSV files about the experiment and each simulation during the build.}
}
\value{
Invisibly returns a list containing the original hip object.
}
\description{
Builds a Hi-sAFe simulation or experiment (a group of simulations) - creates the folder structure and input files.
}
\examples{
\dontrun{
# For a single Hi-sAFe simulation
mysim <- define_hisafe(path = "./simulations", latitude = 30)

# Building the simulation folder structure & files:
build_hisafe(mysim)

# Once a group Hi-sAFe simulations (experiment) is defined:
myexp <- define_hisafe(path = "./simulations", latitude = c(30,60))

# Building the experiment folder structure & files:
build_hisafe(myexp)
}
}
