% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_structure}
\alias{build_structure}
\title{Builds the structure of a Hi-sAFe simulation}
\usage{
build_structure(exp.plan, path, profiles, template)
}
\arguments{
\item{exp.plan}{The exp.plan element of a "hip" object, containing a single row.}

\item{path}{A character string of the path to the simulation folder.}

\item{profiles}{A character vector of export profiles the simulation to export.}

\item{template}{A character string of the path to the Hi-sAFe directory structure/files to use as a template (or one of the strings signaling a default template)}
}
\value{
Invisibly returns a list containing the original hip object and supplied path.
}
\description{
Does the heavy lifting for \code{\link{build_hisafe}}.
}
