% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_structure}
\alias{build_structure}
\title{Builds the structure of a Hi-sAFe simulation}
\usage{
build_structure(
  exp.plan,
  path,
  profiles,
  template,
  files,
  plot.scene,
  summary.files,
  stics.diagnostics
)
}
\arguments{
\item{exp.plan}{The exp.plan element of a "hip" object, containing a single row.}

\item{path}{A character string of the path to the simulation folder.}

\item{profiles}{A character vector of export profiles the simulation to export.}

\item{template}{A character string of the path to the Hi-sAFe directory structure/files to use as a template
(or one of the strings signaling a default template)}

\item{files}{A character string of file types indicating which simulation files to build. Use "all" to write all required simulation files.
Otherwise, select one or more of "sim", "pld", "wth", "tree", "plt", "tec", "par", and "pro".}

\item{plot.scene}{Logical indicating whether \code{\link{plot_hisafe_scene}} should be used to export plots of each scene during the build.}

\item{summary.files}{Logical indicating whether or not to write out summary .CSV files about the experiment and each simulation during the build.}

\item{stics.diagnostics}{Logical indicating whether or not STICS diagnostics files should be exported in the simulation.}
}
\value{
Invisibly returns a list containing the original hip object and supplied path.
}
\description{
Does the heavy lifting for \code{\link{build_hisafe}}.
}
\keyword{internal}
