% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{check_input_values}
\alias{check_input_values}
\title{Check validity of all Hi-sAFe inputs}
\usage{
check_input_values(hip, force)
}
\arguments{
\item{hip}{An object of class "hip".}

\item{force}{Logical indicating wether the supplied values should be forced past the constraint checks. Use \code{TRUE} for development only.}
}
\value{
Produces errors if issues are found. Otherwise, invisibly returns \code{TRUE}.
}
\description{
Checks the validity of all defined inputs against Hi-sAFe model constraints
and constraints found in the template file comments. Errors are generated if issues are found.
Used within \code{\link{define_hisafe}} and \code{\link{define_hisafe_file}}.
}
