% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{copy_hisafe_template}
\alias{copy_hisafe_template}
\title{Copy a Hi-sAFe template to specified location}
\usage{
copy_hisafe_template(template, destination, overwrite = TRUE)
}
\arguments{
\item{template}{A character string of the path to the Hi-sAFe directory structure/files to use as a template
(or one of the strings signaling a default template)
See \code{\link{define_hisafe}} for more details on available default templates.}

\item{path}{A character string of the path to where the template folder should be copied.}
}
\value{
Invisibly returns a logical vector indicating if the attempted fily copy succeeded.
}
\description{
Copies a Hi-sAFe template to specified location.
}
\examples{
\dontrun{
copy_hisafe_template("agroforestry_default", "/Users/myname/Desktop/")
}
}
