% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{copy_hisafe_template}
\alias{copy_hisafe_template}
\title{Copy a Hi-sAFe template to specified location}
\usage{
copy_hisafe_template(template, destination, overwrite = TRUE,
  new.name = NULL)
}
\arguments{
\item{template}{A character string of the path to the Hi-sAFe directory structure/files to use as a template
(or one of the strings signaling a default template)
See \code{\link{define_hisafe}} for more details on available default templates.}

\item{destination}{A character string of the path to where the template folder should be copied.}

\item{overwrite}{A logical indicating whether or not to allow overwriting of an existing folder.}

\item{new.name}{A character string of the a name for the newly copied folder.
If \code{NULL}, the default, then the name will remain the same as the original template folder.}
}
\value{
Invisibly returns a logical vector indicating if the attempted fily copy succeeded.
}
\description{
Copies a Hi-sAFe template to specified location.
}
\examples{
\dontrun{
copy_hisafe_template("agroforestry_default", "/Users/myname/Desktop/")
}
}
\seealso{
Other hisafe helper functions: \code{\link{get_pheno_dates}},
  \code{\link{get_pruning_dates}},
  \code{\link{hip_params}}, \code{\link{hisafe_info}},
  \code{\link{hisafe_profiles}}, \code{\link{hop_filter}},
  \code{\link{hop_merge}}, \code{\link{hop_params}},
  \code{\link{hop_rename}}, \code{\link{is_hip}},
  \code{\link{is_hop}}, \code{\link{join_profiles}},
  \code{\link{profile_check}},
  \code{\link{variable_check}},
  \code{\link{warn_unequal_lengths}},
  \code{\link{which_profiles}}
}
\concept{hisafe helper functions}
