% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define_hisafe}
\alias{define_hisafe}
\title{Define a Hi-sAFe experiment}
\usage{
define_hisafe(factorial = FALSE, ...)
}
\arguments{
\item{factorial}{If \code{FALSE}, the default, then values are recycled (i.e. such as for default behavior of \code{data.frame()}).
If \code{TRUE}, then a factorial experiment is created, in which an ecperiment is defined for each possible combination of supplied values.}

\item{...}{Any suported Hi-sAFe input parameter in the .sim and .pld files can be passed. Parameters should be passed as \code{parameterName = values}. For more information on supported parameters, see _____.}
}
\value{
An object of class \code{hip}. This is a data frame with each row a Hi-sAFe simulation and each column a Hi-sAFe input parameter.
}
\description{
Defines a Hi-sAFe experiment - the input parameters to one or more Hi-sAFe simulations.
}
\details{
It is strongly recommended to name each simulation in your experiment. This can be done via the \code{SimulationName} parameter.
If no names are provided, then \code{define_hisafe} will provide generic names of "Sim_1", "Sim_2", etc.
If \code{factorial = FALSE}, the default, then supply a value of \code{SimulationName} for each experiment.
#' If \code{factorial = TRUE}, then supply a single value of \code{SimulationName} as a prefix.
\code{define_hisafe} will append "_1", "_2", etc. for each of the (unknown number of) factorial experiments.
}
\examples{
\dontrun{
# To define a Hi-sAFe simulation using all default parameter values:
default.exp <- define_hisfafe()

# To define a Hi-sAFe experiment analyzing only variation in latitude:
latitude.exp <- define_hisfafe(SimulationName = c("Lat15", "Lat30", "Lat45", "Lat60"),
                               latitude = seq(15, 60, 15))

# To define a factorial Hi-sAFe experiment analyzing
# simultaneous variation in latitude and tree line orientation:
lat.orient.exp <- define_hisfafe(SimulationName = c("Lat15", "Lat30", "Lat45", "Lat60"),
                                 latitude = seq(15, 60, 15),
                                 treeLineOrientation = c(0,90))
}
}
\seealso{
Other hisafe definition functions: \code{\link{define_hisafe_file}}
}
