% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diag_hisafe}
\alias{diag_hisafe}
\title{Shortcut to Hi-sAFe diagnostics}
\usage{
diag_hisafe(hop, trees = TRUE, plot = TRUE, climate = TRUE,
  cells = TRUE, monthCells = TRUE, annualCells = FALSE, voxels = TRUE,
  ...)
}
\arguments{
\item{hop}{An object of class hop or face.}

\item{trees}{Logical indicating if trees profile diagnostic plots should be made.}

\item{plot}{Logical indicating if plot profile diagnostic plots should be made.}

\item{climate}{Logical indicating if climate profile diagnostic plots should be made.}

\item{cells}{Logical indicating if cells profile diagnostic plots should be made.}

\item{monthCells}{Logical indicating if monthCells profile diagnostic plots should be made.}

\item{annualCells}{Logical indicating if annualCells profile diagnostic plots should be made.}

\item{voxels}{Logical indicating if voxels profile diagnostic plots should be made.}

\item{...}{Other arguments passed to \code{\link{plot_hisafe_ts}}.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Runs the various Hi-sAFe diagnostic functions from a single call.
}
\examples{
\dontrun{
diag_hisafe(myhop)
}
}
