% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diag_hisafe_annualcrop}
\alias{diag_hisafe_annualcrop}
\title{Plot annualcrop diagnostics of Hi-sAFe output}
\usage{
diag_hisafe_annualcrop(hop, output.path = NULL, trees = TRUE,
  canopies = TRUE)
}
\arguments{
\item{hop}{An object of class "hop" or "hop-group" containing output data from one or more Hi-sAFe simulations.}

\item{output.path}{A character stting indicating the path to the directory where plots should be saved. Plots are
saved in a subdirectory within this directory named /annualcrop
If no value is provided, the experiment/simulation path is read from the hop object, and a directory is created there called "analysis/diagnostics".}

\item{trees}{Logical indicating if a point should be plotted at the location of each tree.}

\item{canopies}{Logical indicating if an elipsoid should be plotted representing the size of each tree canopy.}
}
\value{
Invisibly returns a list of \code{ggplot} objects.
}
\description{
Creates a tile plot of every Hi-sAFe annualcrop output variable. All plots are saved as png files to a specifified output path.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe(path = "./")

# You can create tile plots of every annualcrop variable:
diag_hisafe_annualcrop(mydata)
}
}
\seealso{
Other hisafe diagnostic fucntions: \code{\link{diag_hisafe_monthcells}},
  \code{\link{diag_hisafe_ts}}
}
