% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diag_hisafe_monthcells}
\alias{diag_hisafe_monthcells}
\title{Plot monthCells diagnostics of Hi-sAFe output}
\usage{
diag_hisafe_monthcells(data, output.path = "./diagnostics")
}
\arguments{
\item{data}{An object of class \code{hop} or \code{hop-group} containing output data from one or more Hi-sAFe simulations.}

\item{output.path}{A character stting indicating the path to the directory where plots should be saved. Plots are
saved in a subdirectory within this directory named /monthCells/facetScheme.}
}
\value{
A list of \code{ggplot} objects is invisibly returned, grouped into two sublists based on the faceting scheme. The first sublist contains  Year~SimulationName, and the second contains Month~SimulationName.
}
\description{
Creates three tile plots of every Hi-sAFe monthCells output variable, one plot for each possible
\code{plot_hisafe_monthcells} facet scheme. All plots are saved as png files to a specifified output path.
}
\details{
The default data presentation for each of the facet schemes is:
Year~SimulationName plots are for Month==6, every 5 years from 0:max, and all simulations.
Month~SimulationName plots are for Year==median, all 12 months, and all simulations.
Month~Year plots are for all 12 months and every 5 years from 0:max, with each simulation in its own plot.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe_output("MySimulation", "./")

# You can create tile plots of every monthCells variable:
diag_hisafe_monthcells(mydata)
}
}
\seealso{
Other hisafe diagnostic fucntions: \code{\link{diag_hisafe_ts}}
}
