% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{diag_hisafe_ts}
\alias{diag_hisafe_ts}
\title{Plot timeseries diagnostics of Hi-sAFe output}
\usage{
diag_hisafe_ts(data, time.class = "Annual", output.path = "./diagnostics",
  time.lim = NULL, tree.id = NULL)
}
\arguments{
\item{data}{An object of class \code{hop} or \code{hop-group} containing output data from one or more Hi-sAFe simulations.}

\item{time.class}{If 'Annual', annual timeseries are created. If 'Daily', daily timeseries are created.}

\item{output.path}{A character stting indicating the path to the directory where plots should be saved. Plots are
saved in a subdirectory within this directory named by \code{time.class}.}

\item{time.lim}{If time.class is 'Annual', a numeric vector of length two providing the \code{c(minimum, maximum)} of years (since planting) to plot.
If time.class is 'Daily, a character vector of length two providing the \code{c(minimum, maximum)} dates ('yyyy-mm-dd') to plot.
If no input, the full available time range is plotted. Use \code{NA} to refer to the start or end of the simulation.}

\item{tree.id}{A numeric vector indicating the ids of a subset of tree ids to plot. If no input, all trees will be plotted.}
}
\value{
A list of \code{ggplot} objects is invisibly returned. If the data is of class \code{hop-group} and contains
data from more than one Hi-sAFe simulation, the plots will contain multiple lines, colored and labeled by SimName.
If the data contains two more tree ids, the plots will be faceted by tree id.
}
\description{
Plots a daily or annual timeseries of every Hi-sAFe output variable. All plots are saved as
png files to a specifified output path.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe_output("MySimulation", "./")

# You can create an annual timeseries of every variable:
diag_hisafe_ts(mydata)

# For aily timeseries instead:
diag_hisafe_ts(mydata, "Daily")
}
}
