% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diag_hisafe_ts}
\alias{diag_hisafe_ts}
\title{Plot timeseries diagnostics of Hi-sAFe output}
\usage{
diag_hisafe_ts(hop, profile, output.path = NULL, simu.names = "all",
  years = "all", tree.ids = "all", doy.lim = c(1, 366),
  color.palette = NULL, linetype.palette = NULL, aes.cols = list(color =
  "SimulationName", linetype = "SimulationName"), facet.year = TRUE,
  crop.points = FALSE)
}
\arguments{
\item{hop}{An object of class "hop" or "hop-group" containing output data from one or more Hi-sAFe simulations.}

\item{profile}{The profile for which to plot a timeseries. If 'annualtree' or 'annualplot', annual timeseries are created.
If 'trees', 'plot', or 'climate', daily timeseries are created.}

\item{output.path}{A character stting indicating the path to the directory where plots should be saved.
Plots aresaved in a subdirectory within this directory named by \code{profile}.
If no value is provided, the experiment/simulation path is read from the hop object, and a directory is created there called "analysis/diagnostics".}

\item{simu.names}{A character vector of the SimulationNames within \code{hop} to include. Use "all" to include all available values.}

\item{years}{A numeric vector of the years within \code{hop} to include. Use "all" to include all available values.}

\item{tree.ids}{A numeric vector indicating a subset of tree ids to plot. Use "all" to include all available values.}

\item{doy.lim}{A numeric vector of length two providing the \code{c(minimum, maximum)} of julian days to plot. Only applies if \code{profile} is a daily profile.}

\item{color.palette}{A character stirng of hex values or R standard color names defining the color palette to use in plots with multiple simulations.
If \code{NULL}, the default, then the default color palette is a color-blind-friendly color palette. The default supports up to 24 simulations.}

\item{linetype.palette}{A character stirng of values defining the linetype palette to use in plots with multiple simulations.
If \code{NULL}, the default, then solid lines are used for all simulations. The default supports up to 24 simulations.}

\item{aes.cols}{A list with arguments "color" and "linetype" containing character stirngs of the column names to use for plot aesthetics.}

\item{facet.year}{A logical indicating whether, for daily profiles, the plot should be faceted by year. This helps with seeing finer level detail.}

\item{crop.points}{Logical indicating if points should be plotted as well, with point shape desgnating the main crop name.
Only applies when \code{profile} is 'plot' or 'annualplot'.}
}
\value{
Invisibly returns a list of \code{ggplot} objects. If the data is of class "hop-group" and contains
data from more than one Hi-sAFe simulation, the plots will contain multiple lines, colored and labeled by SimulationName.
If the data contains two more tree ids, the plots will be faceted by tree id.
}
\description{
Plots a daily or annual timeseries of every Hi-sAFe output variable in a specified output profile.
All plots are saved as png files to a specifified output path.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe_output("MySimulation", "./")

# You can create an annual timeseries of every variable:
diag_hisafe_ts(mydata)

# For daily timeseries instead:
diag_hisafe_ts(mydata, "daily")
}
}
\seealso{
Other hisafe diagnostic fucntions: \code{\link{diag_hisafe_monthcells}}
}
