% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diag_hisafe_ts}
\alias{diag_hisafe_ts}
\title{Plot timeseries diagnostics of Hi-sAFe output}
\usage{
diag_hisafe_ts(hop, profile, output.path = NULL, ...)
}
\arguments{
\item{hop}{An object of class "hop" or "face" containing output data from one or more Hi-sAFe simulations.}

\item{profile}{The profile for which to plot a timeseries. One of 'trees', 'plot', 'climate', or 'cells'.}

\item{output.path}{A character string indicating the path to the directory where plots should be saved.
Plots aresaved in a subdirectory within this directory named by \code{profile}.
If no value is provided, the experiment/simulation path is read from the hop object, and a directory is created there called "analysis/diagnostics".}

\item{...}{Other arguments passed to \code{\link{plot_hisafe_ts}}.}
}
\value{
Invisibly returns a list of \code{ggplot} objects. If \code{hop} contains
data from more than one Hi-sAFe simulation, the plots will contain multiple lines, colored and labeled by SimulationName.
If the data contains two more tree ids, the plots will be faceted by tree id.
}
\description{
Plots a daily timeseries of every Hi-sAFe output variable in a specified output profile.
All plots are saved as png files to a specifified output path.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe(path = "mydir", simu.names = "MySimulation")

# You can create a daily timeseries of every trees variable:
diag_hisafe_ts(mydata, "trees")
}
}
\seealso{
Other hisafe diagnostic fucntions: \code{\link{diag_hisafe_annualcells}},
  \code{\link{diag_hisafe_cells}},
  \code{\link{diag_hisafe_monthcells}},
  \code{\link{diag_hisafe_voxels}}
}
\concept{hisafe diagnostic fucntions}
