% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diag_hisafe_ts}
\alias{diag_hisafe_ts}
\title{Plot timeseries diagnostics of Hi-sAFe output}
\usage{
diag_hisafe_ts(hop, time.class = "annual", output.path = NULL,
  time.lim = NULL, tree.id = NULL)
}
\arguments{
\item{hop}{An object of class "hop" or "hop-group" containing output data from one or more Hi-sAFe simulations.}

\item{time.class}{If 'annual', the default, annual timeseries are created. If 'daily', daily timeseries are created.}

\item{output.path}{A character stting indicating the path to the directory where plots should be saved. Plots are
saved in a subdirectory within this directory named by \code{time.class}.
If no value is provided, the experiment/simulation path is read from the hop object, and a folder is created there called "diagnostics".}

\item{time.lim}{If time.class is 'annual', the default, a numeric vector of length two providing
the \code{c(minimum, maximum)} of years (since planting) to plot.
If time.class is 'daily', a character vector of length two providing the \code{c(minimum, maximum)} dates ('yyyy-mm-dd') to plot.
If no input, the full available time range is plotted. Use \code{NA} to refer to the start or end of the simulation.}

\item{tree.id}{A numeric vector indicating the ids of a subset of tree ids to plot. If no input, all trees will be plotted.}
}
\value{
Invisibly returns a list of \code{ggplot} objects. If the data is of class "hop-group" and contains
data from more than one Hi-sAFe simulation, the plots will contain multiple lines, colored and labeled by SimulationName.
If the data contains two more tree ids, the plots will be faceted by tree id.
}
\description{
Plots a daily or annual timeseries of every Hi-sAFe output variable. All plots are saved as
png files to a specifified output path.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe_output("MySimulation", "./")

# You can create an annual timeseries of every variable:
diag_hisafe_ts(mydata)

# For daily timeseries instead:
diag_hisafe_ts(mydata, "daily")
}
}
\seealso{
Other hisafe diagnostic fucntions: \code{\link{diag_hisafe_monthcells}}
}
