% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{extract_complete_dates}
\alias{extract_complete_dates}
\title{Extract complete dates}
\usage{
extract_complete_dates(hop, profile, dates)
}
\arguments{
\item{hop}{An object of class hop.}

\item{profile}{Character string of the name of a hop profile.}

\item{dates}{Character vector or Date vector of dates to check.}
}
\description{
Extracts complete dates for \code{\link{hisafe_slice}}
}
\keyword{internal}
