% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{get_light_fluxes}
\alias{get_light_fluxes}
\title{Get light fluxes from a hop object}
\usage{
get_light_fluxes(hop)
}
\arguments{
\item{hop}{An object of class hop or face.}
}
\value{
A tibble with extracted and calculated light fluxes.
}
\description{
Get light fluxes from a hop object.
Used within hisafe cycle functions.
}
\keyword{internal}
