% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{get_template_params}
\alias{get_template_params}
\title{Read all template parameters}
\usage{
get_template_params(template.path)
}
\arguments{
\item{template.path}{A character string of the path to the directory containing the template set of Hi-sAFe simulation folders/files to use.}
}
\value{
A list containing all parameter values and constraints.
}
\description{
Reads all template parameter values and constraints
}
