% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{get_template_params}
\alias{get_template_params}
\title{Read all template parameters}
\usage{
get_template_params(template)
}
\arguments{
\item{template}{A character string of the path to the Hi-sAFe directory structure/files to use as a template
(or one of the strings signaling a default template)}
}
\value{
A list containing all parameter values and constraints.
}
\description{
Reads all template parameter values and constraints
}
