% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_units}
\alias{get_units}
\title{Extract variable units from a hop object}
\usage{
get_units(hop, profile, variable)
}
\arguments{
\item{hop}{An object of class hop.}

\item{profile}{A characgter string of the profile that \code{variable} is from.}

\item{variable}{A character string of the name of the variable to get units for.}
}
\description{
Extracts variable units from a hop object. Use within all hisafe plot functions.
}
