% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{get_water_fluxes}
\alias{get_water_fluxes}
\title{Get water fluxes from a hop object}
\usage{
get_water_fluxes(hop)
}
\arguments{
\item{hop}{An object of class hop or face.}
}
\value{
A tibble with extracted and calculated water fluxes.
}
\description{
Get water fluxes from a hop object.
Used within hisafe cycle functions.
}
