% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hisafe_info}
\alias{hisafe_info}
\title{Display version numbers of Hi-sAFe and Java}
\usage{
hisafe_info(capsis.path)
}
\arguments{
\item{capsis.path}{A character string of the path to the Capsis folder}
}
\value{
Invisibly returns the Hi-sAFe version number
}
\description{
Displays the version numbers of Hi-sAFe and Java.
}
\examples{
\dontrun{
hisafe_info()
}
}
\seealso{
Other hisafe helper functions: 
\code{\link{copy_hisafe_template}()},
\code{\link{get_pheno_dates}()},
\code{\link{get_pruning_dates}()},
\code{\link{hip_params}()},
\code{\link{hisafe_profiles}()},
\code{\link{hop_filter}()},
\code{\link{hop_merge}()},
\code{\link{hop_params}()},
\code{\link{hop_rename}()},
\code{\link{is_face}()},
\code{\link{is_hip}()},
\code{\link{is_hop}()},
\code{\link{join_profiles}()},
\code{\link{make_rel_years}()},
\code{\link{profile_check}()},
\code{\link{variable_check}()},
\code{\link{which_profiles}()}
}
\concept{hisafe helper functions}
