% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hisafe_profiles}
\alias{hisafe_profiles}
\title{Display supported Hi-sAFe output profiles}
\usage{
hisafe_profiles(variable = "names")
}
\value{
Invisibly returns a data frame containing the profiles names and output frequency.
}
\description{
Displays supported Hi-sAFe output profiles and standard output frequency.
}
\examples{
\dontrun{
hisafe_profiles()
}
}
\seealso{
Other hisafe helper functions: \code{\link{hisafe_params}},
  \code{\link{hop_merge}}, \code{\link{simu_rename}}
}
