% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hisafe_profiles}
\alias{hisafe_profiles}
\title{Display supported Hi-sAFe output profiles}
\usage{
hisafe_profiles(variable = "names")
}
\value{
Invisibly returns a data frame containing the profiles names and output frequency.
}
\description{
Displays supported Hi-sAFe output profiles and standard output frequency.
}
\examples{
\dontrun{
hisafe_profiles()
}
}
\seealso{
Other hisafe helper functions: \code{\link{hip_params}},
  \code{\link{hop_filter}}, \code{\link{hop_merge}},
  \code{\link{hop_params}}, \code{\link{hop_rename}},
  \code{\link{is_hip}}, \code{\link{is_hop}},
  \code{\link{profile_check}}, \code{\link{variable_check}}
}
