% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{hisafe_slice}
\alias{hisafe_slice}
\title{Plot a "sliced" profile of the simulated Hi-sAFe scene}
\usage{
hisafe_slice(hop, date, simu.names = "all", tree.ids = "all",
  plot.x = "x", Y = "all", rel.dates = NULL, max.soil.depth = NA,
  vars = list(crown.alpha = "leafArea", trunk.alpha = "carbonLabile",
  crop.alpha = "lai", yield.alpha = "eai", voxel.alpha = "totalTreeRootDensity",
  voxel.border = "cropRootDensity", voxel.L.size = "waterStock", voxel.C.size =
  "totalTreeCoarseRootBiomass", voxel.R.size = "mineralNitrogenStock",
  voxel.L.alpha = "totalTreeWaterUptake", voxel.C.alpha = "fineRootCost",
  voxel.R.alpha = "totalTreeNitrogenUptake"), trees = TRUE, crops = TRUE,
  voxels = TRUE, climate = TRUE, mem.max = 0, plot.rows = 1)
}
\arguments{
\item{hop}{An object of class hop.}

\item{date}{A character string of the date to plot, in the format "YYYY-MM-DD" or of class Date.}

\item{simu.names}{A character vector of the SimulationNames within \code{hop} to include. Use "all" to include all available values.}

\item{tree.ids}{A numeric vector indicating a subset of tree ids to plot. Use "all" to include all available values.}

\item{plot.x}{Either "x" or "y", indicating which axis of the Hi-sAFe scene should be plotted along the x-axis of the plot.}

\item{Y}{A numeric vector indicating a subset of cell locations to include when summarizing cell/voxel data.
If \code{plot.x} is "x", then this refers to the y coordinates of cells. If \code{plot.x} is "y", then this refers to the x coordinates of cells.}

\item{rel.dates}{A character vector (in the format "YYYY-MM-DD") or a vector of class Date of the dates from which to scale \code{variable}.
In the plot, \code{variable} will be scaled to be between the minimum and maximum values of \code{variable} across these dates.}

\item{max.soil.depth}{The maximum depth for which to plot voxel data. To reduce confusion, this value can be supplied as positive or negative.}

\item{vars}{A list of variable names specifying which simulated variables in \code{hop} should be represented by various plot components.
Plot components include:
\itemize{
 \item{"crown.alpha"}{ - transparency of the tree crown}
 \item{"trunk.alpha"}{ - transparency of the tree trunk}
 \item{"crop.alpha"}{ - transparency of the crop}
 \item{"yield.alpha"}{ - transparency of the yield portion of the crop}
 \item{"voxel.alpha"}{ - transparency of the voxel}
 \item{"voxel.border"}{ - thickness of the voxel border}
 \item{"voxel.L.size"}{ - size of the left circle within the voxel}
 \item{"voxel.C.size"}{ - size of the center circle within the voxel}
 \item{"voxel.R.size"}{ - size of the right circle within the voxel}
 \item{"voxel.L.alpha"}{ - transparency of the left circle within the voxel}
 \item{"voxel.C.alpha"}{ - transparency of the center circle within the voxel}
 \item{"voxel.R.alpha"}{ - transparency of the right circle within the voxel}
}
All aesthetics set by \code{vars} are relative to the maximum value in each simulation-year combination.}

\item{trees}{A logical indicating whether or not to plot the trees in the scene.}

\item{crops}{A logical indicating whether or not to plot the crops in the scene.}

\item{voxels}{A logical indicating whether or not to plot the voxels in the scene.}

\item{climate}{A logical indicating whether or not to plot climate aspects (e.g. precipitation) in the scene.}

\item{mem.max}{An integer specifying the maximum number of days into the past to search
for crop/voxel data when no data is available for \code{date} within \code{hop}.}

\item{plot.rows}{An integer specifying \code{nrow} passed to \code{ggplot2::facet_wrap}.}
}
\value{
A ggplot object.
}
\description{
Plots a "sliced" profile of the simulated Hi-sAFe scene. Requires the ggforce package.
}
\examples{
\dontrun{
}
}
\seealso{
Other hisafe visualization functions: \code{\link{hisafe_snapshot}}
}
