% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisafer.R
\docType{package}
\name{hisafer}
\alias{hisafer}
\alias{hisafer-package}
\title{hisafe: A package of tools for interacting with the Hi-sAFe biophysical agroforestry model.}
\description{
The hisafer package provides functions for (1) creating Hi-sAFe parameter files and batches
of runs for simulated experiments, (2) calling Hi-sAFe to run simulations, and (3) reading, analyzing,
and plotting Hi-sAFe output data.
}
