% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hop_merge}
\alias{hop_merge}
\title{Merge multiple hop objects}
\usage{
hop_merge(...)
}
\arguments{
\item{...}{Any number of individual hop objects to merge.}
}
\value{
Returns a hop object.
}
\description{
Merges multiple hop objects, renaming simulation names if there are duplicates
}
\examples{
\dontrun{
new_hop <- hop_merge(hop1, hop2, hop3)
}
}
\seealso{
Other hisafe helper functions: \code{\link{hisafe_params}},
  \code{\link{hisafe_profiles}}, \code{\link{hop_filter}},
  \code{\link{simu_rename}}
}
