% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hop_params}
\alias{hop_params}
\title{Display Hi-sAFe output variables}
\usage{
hop_params(variable = "names", search = FALSE, quiet = FALSE)
}
\arguments{
\item{variable}{A character vector of specific Hi-sAFe output variables of which to display details.
Can also be a regular expression for which to search in Hi-sAFe output variable names.
If "names", the default, then just the names of Hi-sAFe output variables is printed to the console.
If "all", then the names, definitions, and units of all Hi-sAFe output variables is printed.}

\item{search}{Logical indicating whether \code{variable} should be treated as a regular expression and
searched for in the variable names rather than matched literally.}

\item{quiet}{Logical indicating whether or not to supress output printed to console.}
}
\value{
Invisibly returns an alphebetized character vector of the names of Hi-sAFe output variables
}
\description{
Displays Hi-sAFe output variables, their definitions, and thier units.
}
\examples{
\dontrun{
hop_params()                 # just for parameter names
hop_params("carbonBranches") # details of cellWidth parameter
hop_params("all")            # details of all parameters
}
}
\seealso{
Other hisafe helper functions: \code{\link{hip_params}},
  \code{\link{hisafe_profiles}}, \code{\link{hop_filter}},
  \code{\link{hop_merge}}, \code{\link{hop_rename}},
  \code{\link{is_hip}}, \code{\link{is_hop}},
  \code{\link{profile_check}}, \code{\link{variable_check}}
}
