% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hop_rename}
\alias{hop_rename}
\title{Change SimulationNames in a hop object}
\usage{
hop_rename(hop, old.names, new.names)
}
\arguments{
\item{hop}{A object of class "hop".}

\item{old.names}{A character vector of the old SimulationNames to change.}

\item{new.names}{A character vector of the new SimulationNames, in the same order as they apply to \code{old.names}.}
}
\value{
Returns the provided hop object with names changed.
}
\description{
Changes SimulationNames in a hop object.
}
\examples{
\dontrun{
hop_rename(myhop, old.names = c("Sim_1", "Sim_2"), new.names = c("Lat30", "Lat60"))
}
}
\seealso{
Other hisafe helper functions: \code{\link{hisafe_params}},
  \code{\link{hisafe_profiles}}, \code{\link{hop_filter}},
  \code{\link{hop_merge}}
}
