% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{layer_init_params}
\alias{layer_init_params}
\title{Generate soil layer initialization table for define_hisafe}
\usage{
layer_init_params(waterContent = c(0.2, 0.3, 0.3, 0.3, 0.3),
  no3Concentration = c(30, 14, 5, 2, 0), nh4concentration = 0)
}
\arguments{
\item{waterContent}{Parameter of Hi-sAFe soil layer initialization table.}

\item{no3Concentration}{Parameter of Hi-sAFe soil layer initialization table.}

\item{nh4concentration}{Parameter of Hi-sAFe soil layer initialization table.}
}
\value{
A list containing one or more data frames (tibbles).
}
\description{
Generates a soil layer initialization table suitable for passing to \code{\link{define_hisafe}}.
}
\examples{
\dontrun{
layer.init <- layer_init_params()
}
}
\seealso{
Other hisafe definition functions: \code{\link{define_hisafe_file}},
  \code{\link{define_hisafe}}, \code{\link{layer_params}},
  \code{\link{root_init_params}},
  \code{\link{tree_init_params}}
}
