% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{layer_params}
\alias{layer_params}
\title{Generate soil layer table for define_hisafe}
\usage{
layer_params(thick = c(0.4, 0.4, 0.6, 1, 7), sand = c(20.35, 16.8, 11.83,
  18.76, 6.5), clay = c(20.15, 24.95, 25.97, 24.7, 32.35),
  limeStone = c(45.7, 49.1, 47.1, 50.4, 52.1), organicMatter = c(2.42, 1.52,
  1.86, 1.67, 3.25), partSizeSand = 290, stone = 0, stoneType = 6,
  infiltrability = 50)
}
\arguments{
\item{thick}{Parameter of Hi-sAFe soil layer table.}

\item{sand}{Parameter of Hi-sAFe soil layer table.}

\item{clay}{Parameter of Hi-sAFe soil layer table.}

\item{limeStone}{Parameter of Hi-sAFe soil layer table.}

\item{organicMatter}{Parameter of Hi-sAFe soil layer table.}

\item{partSizeSand}{Parameter of Hi-sAFe soil layer table.}

\item{stone}{Parameter of Hi-sAFe soil layer table.}

\item{stoneType}{Parameter of Hi-sAFe soil layer table.}

\item{infiltrability}{Parameter of Hi-sAFe soil layer table.}
}
\value{
A list containing one or more data frames (tibbles).
}
\description{
Generates a soil layer table suitable for passing to \code{\link{define_hisafe}}.
}
\examples{
\dontrun{
layers <- layer_params()
}
}
\seealso{
Other hisafe definition functions: \code{\link{define_hisafe_file}},
  \code{\link{define_hisafe}},
  \code{\link{layer_init_params}},
  \code{\link{root_init_params}},
  \code{\link{tree_init_params}}
}
