% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hisafe_cells}
\alias{plot_hisafe_cells}
\title{Tile plot of Hi-sAFe cells output variable}
\usage{
plot_hisafe_cells(hop, variable, dates)
}
\arguments{
\item{hop}{An object of class \code{hop} or \code{hop-group} containing output data from one or more Hi-sAFe simulations.}

\item{variable}{A character string of the name of the variable to color the tiles.}

\item{dates}{A character vector containing the dates (yyyy-mm-dd) to include.}
}
\value{
A ggplot object.
}
\description{
Plots a tile plot of a single Hi-sAFe cells output variable.
SimulationName is used as the column facet. Date is used as the row facet.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe_output("MySimulation", "./")

# You can create a tile plot of relativeDirectParIncident:
tile.plot <- plot_hisafe_cells(mydata, "relativeDirectParIncident", paste(1998, 6:8, 1, sep = "-"))

# Once you have the plot object, you can display it and save it:
tile.plot
ggplot2::ggsave("tiled_relativeDirectParIncident.png", tile.plot)
}
}
\seealso{
Other hisafe plot functions: \code{\link{plot_hisafe_monthcells}},
  \code{\link{plot_hisafe_ts}},
  \code{\link{plot_hisafe_voxels}}
}
