% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{plot_hisafe_cycle_annual}
\alias{plot_hisafe_cycle_annual}
\title{Plot annual barchart of major cycles}
\usage{
plot_hisafe_cycle_annual(hop, cycle, simu.names = "all",
  tree.ids = "all", year.lim = c(NA, NA), doy.start = 1,
  color.palette = NULL, bar.color = "black",
  crop.names = c("Main crop", "Inter crop"), plot = TRUE)
}
\arguments{
\item{hop}{An object of class hop or face.}

\item{cycle}{One of "carbon", "nitrogen", "water", "light", or "yield".}

\item{simu.names}{A character vector of the SimulationNames within \code{hop} to include. Use "all" to include all available values.}

\item{tree.ids}{A numeric vector indicating a subset of tree ids to plot. Use "all" to include all available values.
This only applies when \code{cycle} is "carbon".}

\item{year.lim}{A numeric vector of length two providing the \code{c(minimum, maximum)} of calendar years to plot.
If no input, the full available time range is plotted. Use \code{NA} to refer to the start or end of the simulation.}

\item{doy.start}{The JulianDay [1-365] on which to start the annual cycle accounting. Use 'sim' to specify the starting JulianDay of the simulation.}

\item{color.palette}{A character string of hex values or R standard color names defining the color palette to use in plots with multiple simulations.
If \code{NULL}, the default, then the default color palette is a color-blind-friendly color palette.}

\item{bar.color}{A hex value or R standard color name defining the color to use for bar plot borders}

\item{crop.names}{A character vector of length 2 containing the names to use in the legend for the mainCrop and interCrop of Hi-sAFe, in that order.}

\item{plot}{If \code{TRUE}, the default, a ggplot object is returned. If \code{FALSE}, the data that would create the plot is returned.}
}
\value{
If \code{plot = TRUE}, returns a ggplot object. If \code{plot = FALSE}, returns the data that would create the plot.
If \code{hop} contains more than one simulation, the plot will be faceted by SimulationName.
}
\description{
Plots an annual barchart of tree carbon pools, water fluxes, nitrogen fluxes, or light capture.
}
\details{
Detailed description of the flux components of the nitrogen and water cycles:

NITROGEN
\itemize{
 \item{"Uptake - Trees"}{ - nitrogen uptake by trees}
 \item{"Uptake - Inter crop"}{ - nitrogen uptake by main crop}
 \item{"Uptake - Main crop"}{ - nitrogen uptake by inter crop}
 \item{"Gaseous losses"}{ - gaseous nitrogen losses via denitrification & volatilization of mineral/organic fertilizer inputs}
 \item{"Run-off"}{ - nitrogen contained in rain water (wet deposition) that runs off the scene}
 \item{"Leaching"}{ - nitrate leaching via the (1) bottom of the scene, (2) artificial drainage pipes, and (3) losses to the water table when the nirate concentration in the water table is lower than the nitrate concentration of voxels that it saturates}
 \item{"Fertilization"}{ - mineral nitrogen added by both mineral & organic fertilizers, plus organic nitrogen added by organic fetilizers}
 \item{"Irrigation"}{ - nitrogen added via irrigation water}
 \item{"Deposition"}{ - nitrogen added via wet deposition}
 \item{"Fixation"}{ - nitrogen added via fixation by crops}
 \item{"Litter"}{ - nitrogen added to soil via aerial and root litter from both trees and crops}
 \item{"Water table"}{ - nitrogen added by the water table when the nirate concentration in the water table is higher than the nitrate concentration of voxels that it saturates}
}

WATER
\itemize{
 \item{"Uptake - Trees"}{ - water uptake by trees}
 \item{"Uptake - Inter crop"}{ - water uptake by main crop}
 \item{"Uptake - Main crop"}{ - water uptake by inter crop}
 \item{"Interception"}{ - rain water intercepted by both tree and crop canopies (and then subsequently evaporated off)}
 \item{"Run-off"}{ - rain & irrigation water that runs off the scene (including both the "surface" run-off associated with soil surface conditions plus the "overflow" runoff associated with saturation of the top soil layer and lack of infiltribility)}
 \item{"Soil evaporation"}{ - water evaporated from (1) soil and (2) intercepted rain held in litter/mulch on soil surface}
 \item{"Drainage"}{ - water drainage (1) out of the bottom of the scene and (2) into artificial drainage pipes}
 \item{"Irrigation"}{ - water added via irrigation}
 \item{"Water table"}{ - water added to soil when the water table saturates voxels (the water added is the difference between field capacity and the current water content of the voxel)}
 \item{"Precipitation"}{ - precipitation (rain & snow)}
}
}
\examples{
\dontrun{
# To plot the simulated water cycle:
water.plot <- plot_hisafe_cycle_annual(myhop, "water")

# Once you have the plot object, you can display it and save it:
water.plot
ggsave_fitmax("water_cycle.png", water.plot)
}
}
\seealso{
Other hisafe analysis functions: \code{\link{LER_summary}},
  \code{\link{LER}}, \code{\link{analyze_hisafe}},
  \code{\link{cycle_summary}}, \code{\link{hisafe_budget}},
  \code{\link{plot_hisafe_cycle_daily}},
  \code{\link{stics_budget_comp}}, \code{\link{write_hop}}
}
\concept{hisafe analysis functions}
