% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hisafe_monthcells}
\alias{plot_hisafe_monthcells}
\title{Tile plot of Hi-sAFe monthCells output variable}
\usage{
plot_hisafe_monthcells(hop, variable, colfacet = "SimulationName",
  rowfacet = "Year", sim.names = "all", years = seq(0, 40, 5),
  months = 6)
}
\arguments{
\item{hop}{An object of class \code{hop} or \code{hop-group} containing output data from one or more Hi-sAFe simulations.}

\item{variable}{A character string of the name of the variable to color the tiles.}

\item{colfacet}{One of "Year", "Month", or "SimulationName", indicating which variable to use for column faceting.}

\item{rowfacet}{One of "Year", "Month", or "SimulationName", indicating which variable to use for row faceting.}

\item{sim.names}{A character string containing the SimulationNames to include. Use "all" to include all available values.}

\item{years}{A numeric vector containing the years to include. Use "all" to include all available values.}

\item{months}{A numeric vector containing the months to include. Use "all" to include all available values.}
}
\value{
Returns a ggplot object.
}
\description{
Plots a tile plot of a single Hi-sAFe monthCells output variable.
If the hop objectcontains the "trees" profile with crown diameter variables, the tree locations and crown outlines are also drawn.
}
\details{
This function is very picky! You can only facet by two of the three manipulable variables: SimulationName, Year, Month.
You must ensure that the one varibale not used for faceting is fixed at a single value.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe_output("MySimulation", "./")

# You can create a tile plot of monthDirectParIncident:
tile.plot <- plot_hisafe_monthCells(mydata, "monthDirectParIncident")

# The default settings use data from June and facet by Year and Simulation Name.
# If you instead want to just look at Year 20 and facet by Month and SimulationName:
tile.plot2 <- plot_hisafe_monthCells(mydata, "monthDirectParIncident",
                                     rowfacet = "SimulationName",
                                     colfacet = "Month",
                                     sim.names = "all",
                                     years = 20,
                                     months = 1:12)

# Once you have the plot object, you can display it and save it:
tile.plot2
ggplot2::ggsave("tiled_monthDirectParIncident.png", tile.plot2)
}
}
\seealso{
Other hisafe plot functions: \code{\link{plot_hisafe_cells}},
  \code{\link{plot_hisafe_scene}},
  \code{\link{plot_hisafe_ts}},
  \code{\link{plot_hisafe_voxels}}
}
