% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_hisafe_ts}
\alias{plot_hisafe_ts}
\title{Plot timeseries of Hi-sAFe output variable}
\usage{
plot_hisafe_ts(data, variable, time.class = "annual", time.lim = NULL,
  tree.id = NULL)
}
\arguments{
\item{data}{An object of class \code{hop} or \code{hop-group} containing output data from one or more Hi-sAFe simulations.}

\item{variable}{A character string of the name of the variable to plot.}

\item{time.class}{If 'annual', the default, an annual timeseries is created. If 'daily', a daily timeseries is created.}

\item{time.lim}{If time.class is 'annual', the default, a numeric vector of length two providing the \code{c(minimum, maximum)} of years (since planting) to plot.
If time.class is 'daily', a character vector of length two providing the \code{c(minimum, maximum)} dates ('yyyy-mm-dd') to plot.
If no input, the full available time range is plotted. Use \code{NA} to refer to the start or end of the simulation.}

\item{tree.id}{A numeric vector indicating the ids of a subset of tree ids to plot. If no input, all trees will be plotted.}
}
\value{
A ggplot object. If the data is of class \code{hop-group} and contains data from more than one
Hi-sAFe simulation, the plot will contain multiple lines, colored and labeled by SimulationName. If the data
contains two more tree ids, the plot will be faceted by tree id.
}
\description{
Plots a daily or annual timeseries of a single Hi-sAFe output variable.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe_output("MySimulation", "./")

# You can create an annual timeseries of carbonCoarseRoots:
annual.plot <- plot_hisafe_ts(mydata, "carbonCoarseRoots")

# For a daily timeseries instead:
daily.plot <- plot_hisafe_ts(mydata, "carbonCoarseRoots", "daily")

# Once you have the plot object, you can display it and save it:
annual.plot
ggplot2::ggsave("annual_carbonCoarseRoots.png", annual.plot)
}
}
\seealso{
Other hisafe plot functions: \code{\link{plot_hisafe_cells}},
  \code{\link{plot_hisafe_monthcells}},
  \code{\link{plot_hisafe_voxels}}
}
