% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hisafe_ts}
\alias{plot_hisafe_ts}
\title{Plot timeseries of Hi-sAFe output variable}
\usage{
plot_hisafe_ts(hop, variables, profile, cumulative = FALSE,
  simu.names = "all", years = "all", tree.ids = "all",
  date.min = NA, date.max = NA, doy.lim = c(1, 365),
  color.palette = NA, linetype.palette = NA, aes.cols = list(color =
  "SimulationName", linetype = "SimulationName"), facet.simu = FALSE,
  facet.year = FALSE, facet.crop = FALSE, intercrop = TRUE,
  crop.points = FALSE, plot = TRUE, save = FALSE,
  output.path = NULL)
}
\arguments{
\item{hop}{An object of class hop.}

\item{variables}{A character vector of the names of the variables to plot.
More than one variable name can be passed, but all variables must be from the same \code{profile}.}

\item{profile}{A character string of the profile for which to plot a timeseries. One of 'trees', 'plot', 'climate', or 'cells'.}

\item{cumulative}{Logical indicating wheter \code{variables} should be cumulated before plotting.}

\item{simu.names}{A character vector of the SimulationNames within \code{hop} to include. Use "all" to include all available values.}

\item{years}{A numeric vector of the years within \code{hop} to include. Use "all" to include all available values.}

\item{tree.ids}{A numeric vector indicating a subset of tree ids to plot. Use "all" to include all available values.}

\item{date.min}{A character string of the minimum date to keep, in the format "YYYY-MM-DD" or of class Date.
If NA, the minimum date in the output data is used.}

\item{date.max}{A character string of the maximum date to keep, in the format "YYYY-MM-DD" or of class Date.
If NA, the maximum date in the output data is used.}

\item{doy.lim}{A numeric vector of length two providing the \code{c(minimum, maximum)} of julian days to plot. Only applies if \code{profile} is a daily profile.}

\item{color.palette}{A character string of hex values or R standard color names defining the color palette to use in plots with multiple simulations.
If \code{NA}, the default, then the default color palette is a color-blind-friendly color palette. The default supports up to 48 simulations.}

\item{linetype.palette}{A character string of values defining the linetype palette to use in plots with multiple simulations.
If \code{NA}, the default, then solid lines are used for all simulations. The default supports up to 48 simulations.}

\item{aes.cols}{A list with arguments "color" and "linetype" containing character strings of the column names to use for plot aesthetics.}

\item{facet.simu}{A logical indicating whether the plot should be faceted by SimulationName This helps when values among simulations are overplotted.}

\item{facet.year}{A logical indicating whether the plot should be faceted by year. This helps with seeing finer level detail.}

\item{facet.crop}{A logical indicating whether the plot should be faceted by cropType (mainCrop vs. interCrop). Only applies when \code{profile} is 'cells'.}

\item{intercrop}{A logical indicating whether the plot should include the interCrop cropType. Only applies when \code{profile} is 'cells'.}

\item{crop.points}{Logical indicating if points should be plotted as well, with point shape desgnating the main crop name.
Only applies when \code{profile} is 'plot'.}

\item{plot}{If \code{TRUE}, the default, a ggplot object is returned. If \code{FALSE}, the data that would create the plot is returned.}

\item{save}{If \code{FALSE}, the default, a ggplot object is returned.
If \code{TRUE}, a ggplot object is returned invisibly and the plot is saved to \code{output.path}.}

\item{output.path}{A character string indicating the path to the directory where plots should be saved.
If \code{NULL}, the experiment/simulation path is read from the hop object, and a directory is created there called "analysis/misc".
The tables and plots will be saved in this directory.}
}
\value{
If \code{plot = TRUE}, returns a ggplot object, otherwise the data that would create the plot is returned.
If the data contains two more tree ids, the plot will be faceted by idTree.
Otherwise, if \code{facet.year = TRUE}, plots of daily profiles will be faceted by year.
If more than one value is passed to \code{variable}, then one plot will be created for each variable and combined using \code{cowplot::plot_grid}.
}
\description{
Plots a daily timeseries of a single Hi-sAFe output variable.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe(path = "./")

# For a daily timeseries:
daily.plot <- plot_hisafe_ts(mydata, "carbonCoarseRoots", "trees")

# Once you have the plot object, you can display it and save it:
daily.plot
ggplot2::ggsave("carbonCoarseRoots.png", daily.plot)
}
}
\seealso{
Other hisafe plot functions: \code{\link{plot_hisafe_annualcells}},
  \code{\link{plot_hisafe_bg}},
  \code{\link{plot_hisafe_cells}},
  \code{\link{plot_hisafe_monthcells}},
  \code{\link{plot_hisafe_scene}},
  \code{\link{plot_hisafe_voxels}}
}
\concept{hisafe plot functions}
