% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hisafe_ts}
\alias{plot_hisafe_ts}
\title{Plot timeseries of Hi-sAFe output variable}
\usage{
plot_hisafe_ts(hop, variable, profile, time.lim = NULL, tree.id = NULL,
  color.palette = NULL, linetype.palette = NULL, aes.cols = list(color =
  NULL, linetype = NULL), plot = TRUE)
}
\arguments{
\item{hop}{An object of class hop.}

\item{variable}{A character string of the name of the variable to plot.}

\item{profile}{The profile for which to plot a timeseries. If 'annualtree' or 'annualplot', annual timeseries are created.
If 'trees', 'plot', or 'climate', daily timeseries are created.}

\item{time.lim}{If profile is an annual profile, a numeric vector of length two providing the \code{c(minimum, maximum)} of years to plot.
If profile is daily profile, a character vector of length two providing the \code{c(minimum, maximum)} dates ('yyyy-mm-dd') to plot.
If no input, the full available time range is plotted. Use \code{NA} to refer to the start or end of the simulation.}

\item{tree.id}{A numeric vector indicating the ids of a subset of tree ids to plot. If no input, all trees will be plotted.}

\item{color.palette}{A character stirng of hex values or R standard color names defining the color palette to use in plots with multiple simulations.
If \code{NULL}, the default, then the default color palette is a color-blind-friendly color palette. The default supports up to 24 simulations.}

\item{linetype.palette}{A character stirng of values defining the linetype palette to use in plots with multiple simulations.
If \code{NULL}, the default, then solid lines are used for all simulations. The default supports up to 24 simulations.}

\item{aes.cols}{A list with arguments "color" and "linetype" containing character stirngs of the column names to use for plot aesthetics.
If \code{NULL}, the default, then SimulationName is used for both aesthetics.}

\item{plot}{If \code{TRUE}, the default, a ggplot object is returned. If \code{FALSE}, the data that would create the plot is returned.}
}
\value{
Returns a ggplot object. If \code{hop} contains more than one simulation, the plot will contain multiple lines, colored and labeled by SimulationName.
If the data contains two more tree ids, the plot will be faceted by tree id.
}
\description{
Plots a daily or annual timeseries of a single Hi-sAFe output variable.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe_output("MySimulation", "./")

# You can create an annual timeseries of carbonCoarseRoots:
annual.plot <- plot_hisafe_ts(mydata, "carbonCoarseRoots")

# For a daily timeseries instead:
daily.plot <- plot_hisafe_ts(mydata, "carbonCoarseRoots", "daily")

# Once you have the plot object, you can display it and save it:
annual.plot
ggplot2::ggsave("annual_carbonCoarseRoots.png", annual.plot)
}
}
\seealso{
Other hisafe plot functions: \code{\link{plot_hisafe_cells}},
  \code{\link{plot_hisafe_monthcells}},
  \code{\link{plot_hisafe_scene}},
  \code{\link{plot_hisafe_voxels}}
}
