% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hisafe_voxels}
\alias{plot_hisafe_voxels}
\title{Tile plot of Hi-sAFe voxels output variable}
\usage{
plot_hisafe_voxels(hop, variable, date.min = NA, date.max = NA,
  simu.names = "all", X = NA, Y = NA, Z = NA, summarize.by = "z",
  facet.simu = TRUE, facet.z = FALSE, vline.dates = NA,
  plot = TRUE)
}
\arguments{
\item{hop}{An object of class hop.}

\item{variable}{A character string of the name of the variable to plot.}

\item{date.min}{A character vector containing the minimum date (yyyy-mm-dd) to include.
If \code{NA}, the default, than the minimum value in the voxels profile is used}

\item{date.max}{A character vector containing the minimum date (yyyy-mm-dd) to include.
If \code{NA}, the default, than the maximum value in the voxels profile is used}

\item{simu.names}{A character string containing the SimulationNames to include. Use "all" to include all available values.}

\item{X}{A numeric vector of the x values of the voxels to include. If \code{NA}, the default, then all x values are used.}

\item{Y}{A numeric vector of the y values of the voxels to include. If \code{NA}, the default, then all y values are used.}

\item{Z}{A numeric vector of the z values of the voxels to include. If \code{NA}, the default, then all z values are used.}

\item{summarize.by}{One of 'x', 'y', or 'z', indicating  an axis over which to average voxel values.
If \code{NA}, the default, then no averaging is done and each voxel is plotted as its own line.}

\item{vline.dates}{A character vector of dates (yyyy-mm-dd) at which to plot dashed vertical reference lines.
If \code{NA}, the default, then no reference lines are drawn.}

\item{plot}{If \code{TRUE}, the default, a ggplot object is returned. If \code{FALSE}, the data that would create the plot is returned.}
}
\value{
Returns ggplot object.
}
\description{
Plots a tile plot of a single Hi-sAFe voxels output variable.
If a single date is provided, SimulationName is used as the column facet. Otherwise, Date is used as the column facet.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe(path = "./")

# You can create a tile plot of waterAvailable:
tile.plot <- plot_hisafe_voxels(mydata, "waterAvailable", paste(1998, 6:8, 1, sep = "-"))

# Once you have the plot object, you can display it and save it:
tile.plot
ggsave_fitmax("tiled_waterAvailable.png", tile.plot)
}
}
\seealso{
Other hisafe plot functions: \code{\link{plot_hisafe_annualcells}},
  \code{\link{plot_hisafe_bg}},
  \code{\link{plot_hisafe_cells}},
  \code{\link{plot_hisafe_monthcells}},
  \code{\link{plot_hisafe_scene}},
  \code{\link{plot_hisafe_ts}}
}
\concept{hisafe plot functions}
