% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{profile_check}
\alias{profile_check}
\title{Check for existiance of profiles in a hop object}
\usage{
profile_check(hop, profiles, error = FALSE)
}
\arguments{
\item{hop}{An object of class hop or face.}

\item{profiles}{A character vector of the names of the profiles to check for.}

\item{error}{Logical indicating whehter or not an error should be thrown if any profiles in \code{profiles} are not found.}
}
\value{
A logical vector the same length as \code{profiles} indicating whether each profile is found in \code{hop}.
}
\description{
Checks for existiance of profiles in a hop object
}
\examples{
\dontrun{
profile_check(myhop, "voxels")
}
}
\seealso{
Other hisafe helper functions: \code{\link{copy_hisafe_template}},
  \code{\link{get_pheno_dates}},
  \code{\link{get_pruning_dates}},
  \code{\link{hip_params}}, \code{\link{hisafe_info}},
  \code{\link{hisafe_profiles}}, \code{\link{hop_filter}},
  \code{\link{hop_merge}}, \code{\link{hop_params}},
  \code{\link{hop_rename}}, \code{\link{is_face}},
  \code{\link{is_hip}}, \code{\link{is_hop}},
  \code{\link{join_profiles}},
  \code{\link{variable_check}},
  \code{\link{which_profiles}}
}
\concept{hisafe helper functions}
