% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_hisafe}
\alias{read_hisafe}
\title{Read output from a single Hi-sAFe simulation}
\usage{
read_hisafe(hip = NULL, simu.name = NULL, path = NULL, profiles = "all",
  show.progress = TRUE)
}
\arguments{
\item{hip}{An object of class "hip". To create a hip object see \code{\link{define_hisafe}}.
Cannot provided both \code{hip} and \code{simu.name}.
If \code{hip} is not provided, then \code{path} is required and the input data for the experiment is read from the experiment
summary .csv file created when building the experiment.}

\item{simu.name}{The \code{SimulationName} of the Hi-sAFe simulation to read. This must be the same as the name of the Hi-sAFe simulation folder.
Cannot provided both \code{hip} and \code{simu.name}.}

\item{path}{A character string of the path to the directory containing the Hi-sAFe simulation folder.
If \code{hip} is not provided, then \code{path} is required. If both \code{hip} and \code{path} are provided, \code{path} is used.}

\item{profiles}{A character vector of the names of Hi-sAFe output profiles to read.
If "all" the default, reads all supported Hi-sAFe output profiles. For currently supported profiles see: \code{\link{hisafe_profiles}}}

\item{show.progress}{Logical indicating whether progress messsages should be printed to the console.}
}
\value{
An object of class "hop". This is a list of 13 data frames (tibbles):
\itemize{
 \item{annualtree}
 \item{annualcrop}
 \item{annualplot}
 \item{trees}
 \item{plot}
 \item{climate}
 \item{roots}
 \item{monthCells}
 \item{cells}
 \item{voxels}
 \item{variables}{ - variable descriptions and units from all profiles}
 \item{exp.plan}{ - the exp.plan of the hip object that generated the simulation}
 \item{path}{ - the path to the simulation folder}
}
}
\description{
Reads the designated output profiles from a single Hi-sAFe simulation.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe(mysimhip)

# If only the annual tree data is required:
mytreedata <- read_hisafe(mysimhip, profiles = "annualtree")
}
}
