% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{read_hisafe}
\alias{read_hisafe}
\title{Read output from a single Hi-sAFe simulation}
\usage{
read_hisafe(simu.name, folder, profiles = c("annualtree", "annualplot",
  "trees", "plot", "monthCells", "cells", "voxels", "climate"))
}
\arguments{
\item{simu.name}{The \code{SimulationName} of the Hi-sAFe simulation. This must be the same as the name of the Hi-sAFe simulation folder.}

\item{folder}{A character string of the path to the directory containing the Hi-sAFe simulation folder (which contains the standard subdirectory with the output)}

\item{profiles}{A character vector of the names of Hi-sAFe output profiles to read. Defaults to reading the core Hi-sAFe output profiles.}
}
\value{
An object of class \code{hop}. This is a list of 6 data frames:
\code{annual}, \code{daily}, \code{monthCells} (monthly cell data), \code{cells}, \code{voxels}, and \code{variables} (variable descriptions and units).
}
\description{
Reads the designated output profiles from a single Hi-sAFe simulation.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe("MySimulation", "./")

# If only the annual tree data is required:
mytreedata <- read_hisafe("MySimulation", "./", profiles = "annualtree")
}
}
