% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_hisafe}
\alias{read_hisafe}
\title{Read output from a single Hi-sAFe simulation}
\usage{
read_hisafe(hip = NULL, simu.name = NULL, path = NULL, profiles = "all")
}
\arguments{
\item{hip}{An object of class \code{hip}. To create a hip object see \code{\link{define_exp}}.
If object contains a \code{path} value, then this can be used without providing \code{path} directly to \code{read_hisafe}.
Cannot provided both \code{hip} and \code{simu.name}.
If \code{hip} is not provided, then \code{path} is required and the input data for the experiment is read from the experiment
summary .csv file created when building the experiment.}

\item{simu.name}{The \code{SimulationName} of the Hi-sAFe simulation to read. This must be the same as the name of the Hi-sAFe simulation folder.
Cannot provided both \code{hip} and \code{simu.name}.}

\item{path}{A character string of the path to the directory containing the Hi-sAFe simulation folder
(which contains the standard subdirectory with the output).
If \code{hip} is not provided, then \code{path} is required. If both \code{hip} and \code{path} are provided, \code{path} is used.}

\item{profiles}{A character vector of the names of Hi-sAFe output profiles to read.
Defaults to reading all supported Hi-sAFe output profiles via "all". For currently supported profiles see: \code{\link{hisafe_profiles}}}
}
\value{
An object of class \code{hop}. This is a list of 10 data frames (tibbles):
\code{annual} (includes data from annualtree and annualplot profiles), \code{daily} (includes data from trees, plot, and climate profiles), \code{annualcrop},
\code{roots}, \code{monthCells}, \code{cells}, \code{voxels}, \code{variables} (variable descriptions and units from all profiles),
\code{inputs} (the hip object that generated the simulation), and \code{path} (the path to the simulation folder).
}
\description{
Reads the designated output profiles from a single Hi-sAFe simulation.
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
mydata <- read_hisafe("MySimulation", "./")

# If only the annual tree data is required:
mytreedata <- read_hisafe("MySimulation", "./", profiles = "annualtree")
}
}
