% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_hisafe}
\alias{read_hisafe}
\title{Read output from one or more Hi-sAFe simulations}
\usage{
read_hisafe(hip = NULL, path = NULL, simu.names = "all",
  profiles = "all", show.progress = TRUE, read.inputs = TRUE,
  max.size = 300, date.min = NA, date.max = NA, dates = NULL)
}
\arguments{
\item{hip}{An object of class "hip". To create a hip object see \code{\link{define_hisafe}}.
If \code{hip} is not provided, then \code{path} is required and the input data for the experiment is read from the experiment
summary .csv file created when building the experiment.}

\item{path}{A character string of the path to the directory containing the Hi-sAFe simulation folders.
If \code{hip} is not provided, then \code{path} is required. If \code{hip} is provided, \code{path} is ignored.}

\item{simu.names}{Names of the simulations to read. If "all", the default, then all simulations in the folder are read.}

\item{profiles}{A character vector of the names of Hi-sAFe output profiles to read.
#' If "all" the default, reads all supported Hi-sAFe output profiles. For currently supported profiles see: \code{\link{hisafe_profiles}}}

\item{show.progress}{Logical indicating whether progress messsages should be printed to the console.}

\item{read.inputs}{Logical indicating whether data should be read from the .PLD and .SIM files for the tree.info and plot.info data slots.
Setting this to \code{FALSE} severly limits hisafer's plotting/analysis capacity.
Should only be set to \code{FALSE} if input files are corrupted.}

\item{max.size}{The maximum file size (MB) that should be read. Files larger than this value will be ignored, with a warning.}

\item{date.min}{A character string of the minimum date to keep, in the format "YYYY-MM-DD".
If NA, the minimum date in the output data is used. Only used if \code{dates} is \code{NULL}.}

\item{date.max}{A character string of the maximum date to keep, in the format "YYYY-MM-DD".
If NA, the maximum date in the output data is used. Only used if \code{dates} is \code{NULL}.}

\item{dates}{A character vector (in the format "YYYY-MM-DD") or a vector of class Date of the dates to keep.
If \code{NULL}, then \code{date.max} and \code{date.min} are used instad.}
}
\value{
An object of class "hop". This is a list of 13 data frames (tibbles):
\itemize{
 \item{trees}
 \item{plot}
 \item{climate}
 \item{cells}
 \item{monthCells}
 \item{annualCells}
 \item{voxels}
 \item{plot.info}{ - plot geometry data for each simulation}
 \item{tree.info}{ - tree species and location data for each simulation}
 \item{exp.plan}{ - the exp.plan of the hip object that generated the simulations}
 \item{metadata}{ - the metadata of each simulations (path to simulation folders, model versions, run date , run duration)}
 \item{exp.path}{ - the path to the experiment folder}
}
}
\description{
Reads the designated output profiles from one or more Hi-sAFe simulations
}
\examples{
\dontrun{
# Reading in Hi-sAFe simulation:
myexp <- read_hisafe(myhip)

# If only the tree data is required:
mytreeexp <- read_hisafe(myhip, profiles = "trees")
}
}
