% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_hisafe_example}
\alias{read_hisafe_example}
\title{Read example Hi-sAFe experiment output}
\usage{
read_hisafe_example(profiles = c("annualplot", "annualtree", "annualcrop",
  "plot", "trees", "climate", "monthCells"), show.progress = TRUE)
}
\arguments{
\item{profiles}{A character vector of the names of Hi-sAFe output profiles to read.
Defaults to reading all supported Hi-sAFe output profiles via "all". For currently supported profiles see: \code{\link{hisafe_profiles}}}

\item{show.progress}{Logical indicating whether progress messsages should be printed to the console.}
}
\value{
An object of class "hop". For more details see \code{?read_hisafe_exp}.
}
\description{
Reads in an example Hi-sAFe experiment. For more details see \code{?read_hisafe_exp}.
}
