% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_hisafe_exp}
\alias{read_hisafe_exp}
\title{Read output from a Hi-sAFe experiment}
\usage{
read_hisafe_exp(hip = NULL, path = NULL, simu.names = "all",
  profiles = "all", show.progress = TRUE, max.size = 3e+08)
}
\arguments{
\item{hip}{An object of class "hip". To create a hip object see \code{\link{define_hisafe}}.
If \code{hip} is not provided, then \code{path} is required and the input data for the experiment is read from the experiment
summary .csv file created when building the experiment.}

\item{path}{A character string of the path to the directory containing the Hi-sAFe simulation folders.
If \code{hip} is not provided, then \code{path} is required. If \code{hip} is provided, \code{path} is ignored.}

\item{simu.names}{Names of the simulations to read. If "all", the default, then all simulations in the experiment folder are read.}

\item{profiles}{A character vector of the names of Hi-sAFe output profiles to read.
#' If "all" the default, reads all supported Hi-sAFe output profiles. For currently supported profiles see: \code{\link{hisafe_profiles}}}

\item{show.progress}{Logical indicating whether progress messsages should be printed to the console.}

\item{max.size}{The maximum file size (in bytes) that should be read. Files larger than this value will be ignored, with a warning.}
}
\value{
An object of class "hop-group". This is a list of 15 data frames (tibbles):
\itemize{
 \item{annualtree}
 \item{annualcrop}
 \item{annualplot}
 \item{trees}
 \item{plot}
 \item{climate}
 \item{roots}
 \item{monthCells}
 \item{cells}
 \item{voxels}
 \item{variables}{ - variable descriptions and units from all profiles}
 \item{tree.info}{ - tree species and location data}
 \item{exp.plan}{ - the exp.plan of the hip object that generated the simulation}
 \item{path}{ - the path to the simulation folder}
 \item{exp.path}{ - the path to the experiment folder}
}
}
\description{
Reads the designated output profiles from a Hi-sAFe experiiment (i.e. a group of Hi-sAFe simulations).
}
\examples{
\dontrun{
# After reading in Hi-sAFe simulation data via:
myexp <- read_hisafe_exp(myexphip)

# If only the annual tree data is required:
mytreeexp <- read_hisafe(myexphip, profiles = "annualtree")
}
}
