% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_profile}
\alias{read_profile}
\title{Read a Hi-sAFe output profile}
\usage{
read_profile(profile, path, simu.name, show.progress, max.size)
}
\arguments{
\item{profile}{A character string of the profile name.}

\item{path}{A character string of the path to the folder containing the profiles.}

\item{simu.name}{A character string of the name of the simulation being read.}

\item{show.progress}{Logical indicating whether progress messsages should be printed to the console.}

\item{max.size}{The maximum file size (MB) that should be read. Files larger than this value will be ignored, with a warning.}
}
\value{
A list containing the profile data and the profile variable definitions.
}
\description{
Wrapper for read_hisafe_output_file.
}
\keyword{internal}
