% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{root_init_params}
\alias{root_init_params}
\title{Generate root initialization table for define_hisafe}
\usage{
root_init_params(template, reps = 1, ...)
}
\arguments{
\item{template}{A character string of the path to the directory containing the template set of Hi-sAFe simulation folders/files to use.
See \code{\link{define_hisafe}} for more details.}

\item{...}{Any parameters of Hi-sAFe root initialization table:
 \itemize{
 \item{"shape"}{}
 \item{"repartition"}{}
 \item{"paramShape1"}{}
 \item{"paramShape2"}{}
 \item{"paramShape3"}{}
}}

\item{rep}{Number of times to repeats the rows of the defined table.}
}
\value{
A list containing a Hi-sAFE root initialization table.
}
\description{
Generates a root initialization table suitable for passing to \code{\link{define_hisafe}}.
The output of this function is always passed to \code{\link{define_hisafe}} via the \code{root.initialization} argument
(see example below). Any passed parameters modify the table in the provided template.
}
\examples{
\dontrun{
hip <- define_hisafe(path = getwd(), template = "agroforestry",
                     root.initialization = root_init_params(template    = "agroforestry",
                                                            paramShape1 = 0.5))
}
}
\seealso{
Other hisafe definition functions: 
\code{\link{define_hisafe_file}()},
\code{\link{define_hisafe}()},
\code{\link{layer_init_params}()},
\code{\link{layer_params}()},
\code{\link{tree_init_params}()},
\code{\link{variety_params}()}
}
\concept{hisafe definition functions}
