% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_hisafe}
\alias{run_hisafe}
\title{Run one or more Hi-sAFe simulations}
\usage{
run_hisafe(hip = NULL, path = NULL, simu.names = "all",
  parallel = FALSE, num.cores = NULL, mem.spec = TRUE, quietly = FALSE,
  capsis.path)
}
\arguments{
\item{hip}{An object of class "hip". Required if \code{path} is not provided.}

\item{path}{A character string of the path (relative or absolute) to the directory containing the simulation folders.
Required if \code{hip} is not provided.}

\item{simu.names}{Names of the simulations to run. If "all", the default, then all simulations are run.}

\item{parallel}{Logical, should parallel computing be used.}

\item{num.cores}{Numbers of cores to use in parallel computing.
If not provided, will default to one less than the total number of available cores.}

\item{mem.spec}{An integer indicating the maximum memory use (Mb) permitted for an individual instance of Capsis/Hi-sAFe.
If \code{TRUE}, will default to the system's total memory divided by the number of cores.
If \code{FALSE}, a memory specification will not be made, and Capsis/Hi-sAFe will run with defaults.}

\item{quietly}{Logical indicating whether status messages should printed to the console.}

\item{capsis.path}{A character string of the path (relative or absolute) to the Capsis folder}
}
\value{
Invisibly returns a list of console logs from simulation runs.
}
\description{
Runs one or more Hi-sAFe simulations
}
\examples{
\dontrun{
# Once a group Hi-sAFe simulations (experiment) is defined...
myexp <- define_hisafe(path = "./simulations", exp.name = "lat_exp", latitude = c(30,60))

# ...and built...
build_hisafe(myexp)

# the simulations can all be run:
run_hisafe(myexp)
}
}
