% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_hisafe_exp}
\alias{run_hisafe_exp}
\title{Run a Hi-sAFe experiment}
\usage{
run_hisafe_exp(hip = NULL, path = NULL, simu.names = "all",
  parallel = FALSE, num.cores = NULL,
  capsis.path = "/Applications/Capsis/")
}
\arguments{
\item{hip}{An object of class "hip" with multiple simulations (rows). Required if \code{path} is not provided.}

\item{path}{A character string of the path (relative or absolute) to the experiment folder containing the simulation folders.
Required if \code{hip} is not provided.}

\item{simu.names}{Names of the simulations to run. If "all", the default, then all simulations are run.}

\item{parallel}{Logical, should parallel computing be used.}

\item{num.cores}{Numbers of cores to use in parallel computing. If not provided, will default to one less than the total number of available cores.}

\item{capsis.path}{A character string of the path to the Capsis folder}
}
\value{
Invisibly returns a list of console logs from simulation runs.
}
\description{
Runs a Hi-sAFe experiment (a group of simulations) - calls Hi-sAFe from command line.
}
\examples{
\dontrun{
# Once a group Hi-sAFe simulations (experiment) is defined...
myexp <- define_hisafe(path = "./simulations", latitude = c(30,60))

# ...and built...
build_hisafe_exp(myexp, exp.name = "lat_exp")

# the simulations can all be run:
run_hisafe_exp(myexp)
}
}
\seealso{
Other hisafe run functions: \code{\link{run_hisafe}}
}
