% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visual_legend}
\alias{visual_legend}
\title{Create a legend for hisafe_visual()}
\usage{
visual_legend(hop, vars, cells.var)
}
\arguments{
\item{hop}{An object of class hop.}

\item{vars}{A list of variable names. See \code{\link{hisafe_slice}} for details.}

\item{cells.var}{A character string of the name of the variable ploted by \code{\link{plot_hisafe_cells}}.}
}
\value{
A ggplot object containing the legend.
}
\description{
Creates a legend for hisafe_visual().
}
