% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{warn_unequal_lengths}
\alias{warn_unequal_lengths}
\title{Provide warning if simulation lengths are unequal}
\usage{
warn_unequal_lengths(hop)
}
\arguments{
\item{hop}{An object of class hop or a hop-like object.}
}
\value{
A tibble (data.frame) containing the SimulationName and durations in years.
}
\description{
Provides warning if simulation lengths are unequal.
}
\keyword{internal}
